Obtenir l'ensemble des campagnes attachées à un site

Obtenir l'ensemble des campagnes attachées à un site

get_campaigns(site_code = NULL, opened_at = NULL, closed_at = NULL,
  type = NULL, ...)

Arguments

site_code
character vecteur contenant les codes de sites pour lesquelles ont désire récupérer les campagnes (ex. 134_111_F01)

opened_at
character vecteur contenant les dates de début de campagnes (ex. 2017-01-30)

closed_at
character vecteur contenant les dates de fin de campagnes (ex. 2017-01-30)

type
character vecteur contenant le type de campagnes d'inventaires réalisé (ex. végétation)

...
httr options; arguments de la fonction httr::GET()

Examples

head(get_campaigns())

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 60 x 22
#>       id site_id type    technicians opened_at  closed_at notes created_at     
#>  * <int>   <int> <chr>   <list>      <chr>      <chr>     <lgl> <chr>          
#>  1     1       1 végéta… <chr [2]>   2016-06-23 2016-06-… NA    2018-08-28T15:…
#>  2     2       2 végéta… <chr [2]>   2016-05-20 2016-07-… NA    2018-08-28T15:…
#>  3     3       3 végéta… <chr [3]>   2016-05-17 2016-07-… NA    2018-08-28T15:…
#>  4     4       4 végéta… <chr [2]>   2016-07-11 2016-07-… NA    2018-08-28T15:…
#>  5     5       5 végéta… <chr [2]>   2016-06-21 2016-06-… NA    2018-08-28T15:…
#>  6     6       6 végéta… <chr [2]>   2016-06-14 2016-07-… NA    2018-08-28T15:…
#>  7     7       7 végéta… <chr [2]>   2016-05-26 2016-07-… NA    2018-08-28T15:…
#>  8     8       8 végéta… <chr [2]>   2016-07-05 2016-07-… NA    2018-08-28T15:…
#>  9     9       9 végéta… <chr [2]>   2016-06-03 2016-07-… NA    2018-08-28T15:…
#> 10    10      10 végéta… <chr [2]>   2016-06-28 2016-06-… NA    2018-08-28T15:…
#> # ... with 50 more rows, and 14 more variables: updated_at <chr>,
#> #   efforts <list>, device <lgl>, lures <list>, landmarks <list>, traps <list>,
#> #   environment.id <int>, environment.campaign_id <int>,
#> #   environment.wind <chr>, environment.sky <chr>, environment.temp_c <int>,
#> #   environment.notes <lgl>, environment.created_at <chr>,
#> #   environment.updated_at <chr>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/campaigns?page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 71.7 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 

head(get_campaigns(type="végétation"))

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 26 x 15
#>       id site_id type    technicians opened_at  closed_at notes created_at     
#>  * <int>   <int> <chr>   <list>      <chr>      <chr>     <lgl> <chr>          
#>  1     1       1 végéta… <chr [2]>   2016-06-23 2016-06-… NA    2018-08-28T15:…
#>  2     2       2 végéta… <chr [2]>   2016-05-20 2016-07-… NA    2018-08-28T15:…
#>  3     3       3 végéta… <chr [3]>   2016-05-17 2016-07-… NA    2018-08-28T15:…
#>  4     4       4 végéta… <chr [2]>   2016-07-11 2016-07-… NA    2018-08-28T15:…
#>  5     5       5 végéta… <chr [2]>   2016-06-21 2016-06-… NA    2018-08-28T15:…
#>  6     6       6 végéta… <chr [2]>   2016-06-14 2016-07-… NA    2018-08-28T15:…
#>  7     7       7 végéta… <chr [2]>   2016-05-26 2016-07-… NA    2018-08-28T15:…
#>  8     8       8 végéta… <chr [2]>   2016-07-05 2016-07-… NA    2018-08-28T15:…
#>  9     9       9 végéta… <chr [2]>   2016-06-03 2016-07-… NA    2018-08-28T15:…
#> 10    10      10 végéta… <chr [2]>   2016-06-28 2016-06-… NA    2018-08-28T15:…
#> # ... with 16 more rows, and 7 more variables: updated_at <chr>,
#> #   efforts <list>, environment <lgl>, device <lgl>, lures <list>,
#> #   landmarks <list>, traps <list>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/campaigns?type=v%C3%A9g%C3%A9tation&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 32.3 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#>