Obtenir les informations sur les cellules depuis l'API de coleo

Obtenir les informations sur les cellules depuis l'API de coleo

get_cells(cell_code = NULL, name = NULL, ...)

Arguments

cell_code
character vecteur contenant les identifiants uniques que l'on désire obtenir. Si cell_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.

name
character vecteur contenant les noms de la cellules que l'on veut obtenir.

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_cells(cell_code=c('111_91'))

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 1 x 7
#>      id name     cell_code created_at   updated_at   geom.type geom.coordinates
#> * <int> <chr>    <chr>     <chr>        <chr>        <chr>     <list>          
#> 1    35 La Vall… 111_91    2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/cells?cell_code=111_91&page=0]
#>   Date: 2018-09-26 21:08
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 283 B
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"

head(get_cells())

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 100 x 7
#>       id name    cell_code created_at   updated_at   geom.type geom.coordinates
#>  * <int> <chr>   <chr>     <chr>        <chr>        <chr>     <list>          
#>  1     1 Jamési… 100_111   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 19 × …
#>  2     2 Jamési… 100_155   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  3     3 Jamési… 102_101   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  4     4 Jamési… 102_102   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  5     5 Jamési… 102_104   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  6     6 Témisc… 102_83    2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 18 × …
#>  7     7 Rouyn-… 102_90    2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  8     8 Abitib… 102_92    2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  9     9 Abitib… 102_95    2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#> 10    10 Abitib… 102_96    2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/cells?page=0]
#>   Date: 2018-09-26 21:08
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 32.4 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[2]]
#> $body
#> # A tibble: 100 x 7
#>       id name     cell_code created_at   updated_at  geom.type geom.coordinates
#>  * <int> <chr>    <chr>     <chr>        <chr>       <chr>     <list>          
#>  1   101 La Tuqu… 129_106   2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#>  2   102 Le Fjor… 129_123   2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#>  3   103 Manicou… 129_132   2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#>  4   104 Sept-Ri… 129_146   2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#>  5   105 Papinea… 129_86    2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 25 × …
#>  6   106 Parc du… 129_94    2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#>  7   107 Matawin… 129_97    2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#>  8   108 Mékinac… 130_101   2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#>  9   109 La Tuqu… 130_103   2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#> 10   110 Saguenay 130_117   2018-08-28T… 2018-08-28… Polygon   <dbl [1 × 5 × 2…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/cells?page=1]
#>   Date: 2018-09-26 21:08
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 30.2 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[3]]
#> $body
#> # A tibble: 100 x 7
#>       id name    cell_code created_at   updated_at   geom.type geom.coordinates
#>  * <int> <chr>   <chr>     <chr>        <chr>        <chr>     <list>          
#>  1   201 La Côt… 149_142   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  2   202 Avigno… 150_134   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 24 × …
#>  3   203 Mingan… 150_161   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 13 × …
#>  4   204 Le Roc… 152_141   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  5   205 Parc F… 152_146   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 9 × 2…
#>  6   206 Mingan… 152_152   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  7   207 Bonave… 154_138   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 7 × 2…
#>  8   208 Parc d… 154_155   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#>  9   209 Le Roc… 155_144   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 7 × 2…
#> 10   210 Mingan… 156_155   2018-08-28T… 2018-08-28T… Polygon   <dbl [1 × 5 × 2…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/cells?page=2]
#>   Date: 2018-09-26 21:08
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 38.1 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[4]]
#> $body
#> # A tibble: 100 x 7
#>       id name  cell_code created_at    updated_at    geom.type geom.coordinates
#>  * <int> <lgl> <chr>     <chr>         <chr>         <chr>     <list>          
#>  1   280 NA    110_181   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  2   281 NA    110_95    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 7 × 2…
#>  3   282 NA    110_98    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 11 × …
#>  4   283 NA    111_100   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  5   284 NA    111_102   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  6   285 NA    111_112   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  7   286 NA    111_114   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  8   287 NA    111_131   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 7 × 2…
#>  9   288 NA    111_155   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#> 10   289 NA    111_178   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/cells?page=3]
#>   Date: 2018-09-26 21:08
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 32.3 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[5]]
#> $body
#> # A tibble: 100 x 7
#>       id name  cell_code created_at    updated_at    geom.type geom.coordinates
#>  * <int> <lgl> <chr>     <chr>         <chr>         <chr>     <list>          
#>  1   401 NA    141_95    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 7 × 2…
#>  2   402 NA    142_156   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  3   403 NA    142_174   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  4   404 NA    142_94    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  5   405 NA    143_128   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  6   406 NA    143_94    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  7   407 NA    143_95    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  8   408 NA    143_96    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  9   409 NA    144_102   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#> 10   410 NA    144_116   2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 19 × …
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/cells?page=4]
#>   Date: 2018-09-26 21:08
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 46.5 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[6]]
#> $body
#> # A tibble: 68 x 7
#>       id name  cell_code created_at    updated_at    geom.type geom.coordinates
#>  * <int> <lgl> <chr>     <chr>         <chr>         <chr>     <list>          
#>  1   480 NA    75_137    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  2   481 NA    76_142    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  3   482 NA    76_149    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  4   483 NA    76_167    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  5   484 NA    76_173    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  6   485 NA    77_180    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  7   486 NA    78_155    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  8   487 NA    78_163    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#>  9   488 NA    79_122    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#> 10   489 NA    79_132    2018-08-28T1… 2018-08-28T1… Polygon   <dbl [1 × 5 × 2…
#> # ... with 58 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/cells?page=5]
#>   Date: 2018-09-26 21:08
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 22.4 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#>