Fonction générique pour retirer de l'information depuis l'API de Coléo

Fonction générique pour retirer de l'information depuis l'API de Coléo

get_gen(endpoint = NULL, query = NULL, flatten = TRUE,
  output = "data.frame", ...)

Arguments

endpoint
character désignant le point d'entrée pour le retrait des données. Un point d'entrée peut être vu comme une table de la base de données.

query
list de paramètres à passer avec l'appel sur le endpoint.

flatten
logical aplatir automatiquement un data.frame imbriqués dans un seul data.frame (obsolete si l'objet retourné n'est pas un data.frame)

output
character choix du type d'objet retourné: data.frame, list, json

...
httr options; arguments de la fonction httr::GET()

Value

Retourne un objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Details

Les points d'accès de l'API sont énuméré dans l'environment de coléo, voir print(endpoints)

Examples

resp <- get_gen("/cells")
length(resp) # Nombre de pages retourné par l'appel sur le point d'accès de l'API.

#> [1] 6

str(resp[[1]])

#> List of 2
#>  $ body    :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	100 obs. of  7 variables:
#>   ..$ id              : int [1:100] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ name            : chr [1:100] "Jamésie (A)" "Jamésie (Z)" "Jamésie (B)" "Jamésie (C)" ...
#>   ..$ cell_code       : chr [1:100] "100_111" "100_155" "102_101" "102_102" ...
#>   ..$ created_at      : chr [1:100] "2018-08-28T15:46:49.118Z" "2018-08-28T15:46:49.305Z" "2018-08-28T15:46:49.422Z" "2018-08-28T15:46:49.485Z" ...
#>   ..$ updated_at      : chr [1:100] "2018-08-28T15:46:49.118Z" "2018-08-28T15:46:49.305Z" "2018-08-28T15:46:49.422Z" "2018-08-28T15:46:49.485Z" ...
#>   ..$ geom.type       : chr [1:100] "Polygon" "Polygon" "Polygon" "Polygon" ...
#>   ..$ geom.coordinates:List of 100
#>   .. ..$ : num [1, 1:19, 1:2] -76.4 -76.4 -76.4 -76.4 -76.4 ...
#>   .. ..$ : num [1, 1:5, 1:2] -70.1 -70.3 -70.5 -70.3 -70.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.3 -77.4 -77.6 -77.5 -77.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.2 -77.3 -77.5 -77.3 -77.2 ...
#>   .. ..$ : num [1, 1:5, 1:2] -76.9 -77.1 -77.2 -77.1 -76.9 ...
#>   .. ..$ : num [1, 1:18, 1:2] -79.3 -79.5 -79.6 -79.6 -79.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -78.6 -78.7 -78.8 -78.7 -78.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -78.3 -78.5 -78.6 -78.5 -78.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -78 -78.1 -78.3 -78.2 -78 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.9 -78 -78.2 -78 -77.9 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77 -77.1 -77.3 -77.2 -77 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.9 -75 -75.2 -75.1 -74.9 ...
#>   .. ..$ : num [1, 1:5, 1:2] -79 -79.1 -79.2 -79.1 -79 ...
#>   .. ..$ : num [1, 1:5, 1:2] -78.8 -78.9 -79 -78.9 -78.8 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.4 -77.5 -77.7 -77.5 -77.4 ...
#>   .. ..$ : num [1, 1:46, 1:2] -79.3 -79.3 -79.3 -79.3 -79.3 ...
#>   .. ..$ : num [1, 1:7, 1:2] -79 -79.1 -79.2 -79.3 -79.3 ...
#>   .. ..$ : num [1, 1:11, 1:2] -76.8 -76.9 -76.9 -77 -77 ...
#>   .. ..$ : num [1, 1:5, 1:2] -75.7 -75.9 -76 -75.9 -75.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -78 -78.1 -78.3 -78.2 -78 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.6 -74.7 -74.8 -74.7 -74.6 ...
#>   .. ..$ : num [1, 1:82, 1:2] -79.1 -79.1 -79.1 -79.1 -79.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -78.5 -78.6 -78.7 -78.6 -78.5 ...
#>   .. ..$ : num [1, 1:5, 1:2] -78 -78.2 -78.3 -78.2 -78 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.1 -74.3 -74.4 -74.3 -74.1 ...
#>   .. ..$ : num [1, 1:13, 1:2] -76.7 -76.8 -76.9 -76.9 -76.9 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.8 -74.9 -75 -74.9 -74.8 ...
#>   .. ..$ : num [1, 1:5, 1:2] -76.9 -77.1 -77.2 -77.1 -76.9 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.4 -74.5 -74.7 -74.5 -74.4 ...
#>   .. ..$ : num [1, 1:73, 1:2] -78.6 -78.7 -78.7 -78.7 -78.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -78.3 -78.4 -78.5 -78.4 -78.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.4 -77.5 -77.6 -77.5 -77.4 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.2 -74.3 -74.5 -74.3 -74.2 ...
#>   .. ..$ : num [1, 1:7, 1:2] -73.8 -73.8 -73.9 -74.1 -74 ...
#>   .. ..$ : num [1, 1:9, 1:2] -73 -73.1 -73.3 -73.3 -73.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.1 -77.2 -77.4 -77.3 -77.1 ...
#>   .. ..$ : num [1, 1:9, 1:2] -76.3 -76.4 -76.5 -76.5 -76.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.2 -77.3 -77.5 -77.3 -77.2 ...
#>   .. ..$ : num [1, 1:5, 1:2] -76.7 -76.9 -77 -76.9 -76.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.8 -77.9 -78.1 -78 -77.8 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.6 -74.8 -74.9 -74.8 -74.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.1 -74.3 -74.4 -74.3 -74.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -72.1 -72.3 -72.4 -72.3 -72.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.2 -77.4 -77.5 -77.4 -77.2 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.1 -77.2 -77.4 -77.3 -77.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -76.6 -76.7 -76.8 -76.7 -76.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -75.7 -75.9 -76 -75.9 -75.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -73.3 -73.5 -73.6 -73.5 -73.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -72.9 -73.1 -73.2 -73.1 -72.9 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.6 -74.7 -74.9 -74.7 -74.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -76.5 -76.6 -76.8 -76.6 -76.5 ...
#>   .. ..$ : num [1, 1:5, 1:2] -76.3 -76.4 -76.5 -76.4 -76.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -73.6 -73.7 -73.8 -73.7 -73.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71.7 -71.8 -72 -71.8 -71.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -77.3 -77.4 -77.5 -77.4 -77.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -75.9 -76 -76.2 -76 -75.9 ...
#>   .. ..$ : num [1, 1:9, 1:2] -73.5 -73.7 -73.8 -73.7 -73.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -72.4 -72.5 -72.6 -72.5 -72.4 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71.1 -71.3 -71.4 -71.3 -71.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.4 -74.5 -74.7 -74.6 -74.4 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.2 -74.3 -74.4 -74.3 -74.2 ...
#>   .. ..$ : num [1, 1:5, 1:2] -72.5 -72.6 -72.8 -72.6 -72.5 ...
#>   .. ..$ : num [1, 1:5, 1:2] -76.5 -76.6 -76.8 -76.7 -76.5 ...
#>   .. ..$ : num [1, 1:5, 1:2] -75.7 -75.8 -76 -75.9 -75.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -70.7 -70.8 -71 -70.8 -70.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -67.3 -67.4 -67.6 -67.4 -67.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -76 -76.2 -76.3 -76.2 -76 ...
#>   .. ..$ : num [1, 1:5, 1:2] -75.6 -75.7 -75.8 -75.7 -75.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -72.6 -72.7 -72.9 -72.7 -72.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -73.4 -73.5 -73.6 -73.5 -73.4 ...
#>   .. ..$ : num [1, 1:5, 1:2] -73.1 -73.2 -73.4 -73.2 -73.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71.5 -71.7 -71.8 -71.7 -71.5 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71.4 -71.5 -71.7 -71.5 -71.4 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71 -71.1 -71.3 -71.1 -71 ...
#>   .. ..$ : num [1, 1:7, 1:2] -76.4 -76.6 -76.6 -76.7 -76.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -75.4 -75.5 -75.7 -75.6 -75.4 ...
#>   .. ..$ : num [1, 1:5, 1:2] -75.2 -75.3 -75.4 -75.3 -75.2 ...
#>   .. ..$ : num [1, 1:5, 1:2] -72.6 -72.7 -72.8 -72.7 -72.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -75.6 -75.7 -75.9 -75.8 -75.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.2 -74.3 -74.4 -74.3 -74.2 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71.6 -71.8 -71.9 -71.8 -71.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71.2 -71.4 -71.5 -71.4 -71.2 ...
#>   .. ..$ : num [1, 1:5, 1:2] -67.8 -67.9 -68.1 -67.9 -67.8 ...
#>   .. ..$ : num [1, 1:7, 1:2] -75.8 -75.9 -76.1 -76.1 -76.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -73.7 -73.8 -73.9 -73.8 -73.7 ...
#>   .. ..$ : num [1, 1:5, 1:2] -73.3 -73.4 -73.6 -73.4 -73.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -69.3 -69.4 -69.6 -69.4 -69.3 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71.9 -72 -72.2 -72 -71.9 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.9 -75 -75.1 -75 -74.9 ...
#>   .. ..$ : num [1, 1:5, 1:2] -70.8 -70.9 -71.1 -71 -70.8 ...
#>   .. ..$ : num [1, 1:5, 1:2] -69.4 -69.6 -69.7 -69.6 -69.4 ...
#>   .. ..$ : num [1, 1:5, 1:2] -68.7 -68.8 -69 -68.8 -68.7 ...
#>   .. ..$ : num [1, 1:9, 1:2] -75.7 -75.8 -75.8 -75.8 -75.9 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.2 -74.4 -74.5 -74.4 -74.2 ...
#>   .. ..$ : num [1, 1:5, 1:2] -74.1 -74.2 -74.4 -74.3 -74.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -73.6 -73.8 -73.9 -73.8 -73.6 ...
#>   .. ..$ : num [1, 1:5, 1:2] -72 -72.1 -72.3 -72.1 -72 ...
#>   .. ..$ : num [1, 1:5, 1:2] -71.1 -71.2 -71.4 -71.2 -71.1 ...
#>   .. ..$ : num [1, 1:5, 1:2] -70 -70.1 -70.3 -70.1 -70 ...
#>   .. .. [list output truncated]
#>   ..- attr(*, "n_records")= int 7
#>  $ response:List of 10
#>   ..$ url        : chr "https://coleo.biodiversite-quebec.ca/api/v1/cells?page=0"
#>   ..$ status_code: int 200
#>   ..$ headers    :List of 9
#>   .. ..$ server                     : chr "nginx/1.10.3 (Ubuntu)"
#>   .. ..$ date                       : chr "Wed, 26 Sep 2018 20:58:23 GMT"
#>   .. ..$ content-type               : chr "application/json; charset=utf-8"
#>   .. ..$ content-length             : chr "32401"
#>   .. ..$ connection                 : chr "keep-alive"
#>   .. ..$ x-powered-by               : chr "Express"
#>   .. ..$ access-control-allow-origin: chr "*"
#>   .. ..$ content-range              : chr "items 0-99/568"
#>   .. ..$ etag                       : chr "W/\"7e91-i8/gpctCRfjFVWxDazHvIm58qMw\""
#>   .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
#>   ..$ all_headers:List of 1
#>   .. ..$ :List of 3
#>   .. .. ..$ status : int 200
#>   .. .. ..$ version: chr "HTTP/1.1"
#>   .. .. ..$ headers:List of 9
#>   .. .. .. ..$ server                     : chr "nginx/1.10.3 (Ubuntu)"
#>   .. .. .. ..$ date                       : chr "Wed, 26 Sep 2018 20:58:23 GMT"
#>   .. .. .. ..$ content-type               : chr "application/json; charset=utf-8"
#>   .. .. .. ..$ content-length             : chr "32401"
#>   .. .. .. ..$ connection                 : chr "keep-alive"
#>   .. .. .. ..$ x-powered-by               : chr "Express"
#>   .. .. .. ..$ access-control-allow-origin: chr "*"
#>   .. .. .. ..$ content-range              : chr "items 0-99/568"
#>   .. .. .. ..$ etag                       : chr "W/\"7e91-i8/gpctCRfjFVWxDazHvIm58qMw\""
#>   .. .. .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
#>   ..$ cookies    :'data.frame':	0 obs. of  7 variables:
#>   .. ..$ domain    : logi(0) 
#>   .. ..$ flag      : logi(0) 
#>   .. ..$ path      : logi(0) 
#>   .. ..$ secure    : logi(0) 
#>   .. ..$ expiration:Classes 'POSIXct', 'POSIXt'  num(0) 
#>   .. ..$ name      : logi(0) 
#>   .. ..$ value     : logi(0) 
#>   ..$ content    : raw [1:32401] 5b 7b 22 69 ...
#>   ..$ date       : POSIXct[1:1], format: "2018-09-26 20:58:23"
#>   ..$ times      : Named num [1:6] 0 0.000023 0.000023 0.00008 0.036059 ...
#>   .. ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
#>   ..$ request    :List of 7
#>   .. ..$ method    : chr "GET"
#>   .. ..$ url       : chr "https://coleo.biodiversite-quebec.ca/api/v1/cells?page=0"
#>   .. ..$ headers   : Named chr [1:3] "application/json, text/xml, application/xml, */*" "application/json" "Bearer f8a708d7d83f3c45a477ee3f8722657ff184b9c86c771c1712c98b425afc1b6a"
#>   .. .. ..- attr(*, "names")= chr [1:3] "Accept" "Content-type" "Authorization"
#>   .. ..$ fields    : NULL
#>   .. ..$ options   :List of 2
#>   .. .. ..$ useragent: chr "rcoleo"
#>   .. .. ..$ httpget  : logi TRUE
#>   .. ..$ auth_token: NULL
#>   .. ..$ output    : list()
#>   .. .. ..- attr(*, "class")= chr [1:2] "write_memory" "write_function"
#>   .. ..- attr(*, "class")= chr "request"
#>   ..$ handle     :Class 'curl_handle' <externalptr> 
#>   ..- attr(*, "class")= chr "response"
#>  - attr(*, "class")= chr "getSuccess"

class(resp[[1]])

#> [1] "getSuccess"