Retourne les entrées pour les repères sur le terrain spécifique à une campagne d'inventaires

Retourne les entrées pour les repères sur le terrain spécifique à une campagne d'inventaires

get_landmarks(site_code = NULL, opened_at = NULL, closed_at = NULL,
  type = NULL, ...)

Arguments

site_code
character vecteur contenant les codes de sites pour lesquelles ont désire récupérer les campagnes (ex. 134_111_F01)

opened_at
character vecteur contenant les dates de début de campagnes (ex. 2017-01-30)

closed_at
character vecteur contenant les dates de fin de campagnes (ex. 2017-01-30)

type
character vecteur contenant le type de campagnes d'inventaires réalisé (ex. végétation)

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_landmarks()

#> [[1]]
#> [[1]][[1]]
#> $body
#>    id campaign_id tree_code             taxa_name dbh dbh_unit azimut distance
#> 1   1           2         5        Acer saccharum 380       mm    200       90
#> 2   2           3         1        Acer saccharum 350       mm     21      213
#> 3   3           6         9        Abies balsamea 200       mm    188      276
#> 4   4           7         6        Acer saccharum 680       mm    221      440
#> 5   5           9        57        Abies balsamea 330       mm    274      400
#> 6   6          21         6      Tsuga canadensis 300       mm    279      216
#> 7   7          22         7           Acer rubrum 250       mm    169      230
#> 8   8          23        29    Betula populifolia 100       mm    169      460
#> 9   9          24        29     Betula papyrifera 140       mm    172      114
#> 10 10          25        17        Acer saccharum 260       mm     14      335
#> 11 11           2        19        Acer saccharum 280       mm     41      234
#> 12 12           3         2           Acer rubrum 140       mm    246      290
#> 13 13           6        10        Abies balsamea 170       mm    230      346
#> 14 14           9        37        Abies balsamea 250       mm    174      472
#> 15 15          21        19     Betula papyrifera 150       mm     17      459
#> 16 16          22        19        Acer saccharum 240       mm     25      393
#> 17 17          23        10        Abies balsamea 240       mm     69      800
#> 18 18          24        58     Betula papyrifera 160       mm     21       93
#> 19 19          25         7 Betula alleghaniensis 390       mm    290      560
#> 20 20          41      <NA>                  <NA>   0       mm      0        0
#> 21 21          42      <NA>                  <NA>   0       mm      0        0
#> 22 22          43      <NA>                  <NA>   0       mm      0        0
#> 23 23          44      <NA>                  <NA>   0       mm      0        0
#> 24 24          45      <NA>                  <NA>   0       mm      0        0
#> 25 25          46      <NA>                  <NA>   0       mm      0        0
#> 26 26          47      <NA>                  <NA>   0       mm      0        0
#> 27 27          48      <NA>                  <NA>   0       mm      0        0
#> 28 28          49      <NA>                  <NA>   0       mm      0        0
#> 29 29          50      <NA>                  <NA>   0       mm      0        0
#> 30 30          51      <NA>                  <NA>   0       mm      0        0
#> 31 31          52      <NA>                  <NA>   0       mm      0        0
#> 32 32          53      <NA>                  <NA>   0       mm      0        0
#> 33 33          54      <NA>                  <NA>   0       mm      0        0
#> 34 34          41      <NA>                  <NA>   0       mm      0        0
#> 35 35          42      <NA>                  <NA>   0       mm      0        0
#> 36 36          43      <NA>                  <NA>   0       mm      0        0
#> 37 37          44      <NA>                  <NA>   0       mm      0        0
#> 38 38          45      <NA>                  <NA>   0       mm      0        0
#> 39 39          46      <NA>                  <NA>   0       mm      0        0
#> 40 40          47      <NA>                  <NA>   0       mm      0        0
#> 41 41          48      <NA>                  <NA>   0       mm      0        0
#> 42 42          49      <NA>                  <NA>   0       mm      0        0
#> 43 43          50      <NA>                  <NA>   0       mm      0        0
#> 44 44          51      <NA>                  <NA>   0       mm      0        0
#> 45 45          52      <NA>                  <NA>   0       mm      0        0
#> 46 46          53      <NA>                  <NA>   0       mm      0        0
#> 47 47          54      <NA>                  <NA>   0       mm      0        0
#> 48 48          41      <NA>                  <NA>   0       mm      0        0
#> 49 49          42      <NA>                  <NA>   0       mm      0        0
#> 50 50          43      <NA>                  <NA>   0       mm      0        0
#> 51 51          44      <NA>                  <NA>   0       mm      0        0
#> 52 52          45      <NA>                  <NA>   0       mm      0        0
#> 53 53          46      <NA>                  <NA>   0       mm      0        0
#> 54 54          47      <NA>                  <NA>   0       mm      0        0
#> 55 55          48      <NA>                  <NA>   0       mm      0        0
#> 56 56          49      <NA>                  <NA>   0       mm      0        0
#> 57 57          50      <NA>                  <NA>   0       mm      0        0
#> 58 58          51      <NA>                  <NA>   0       mm      0        0
#> 59 59          52      <NA>                  <NA>   0       mm      0        0
#> 60 60          53      <NA>                  <NA>   0       mm      0        0
#> 61 61          54      <NA>                  <NA>   0       mm      0        0
#>    distance_unit type notes               created_at               updated_at
#> 1             cm both    NA 2018-08-16T17:25:15.164Z 2018-08-16T17:25:15.164Z
#> 2             cm both    NA 2018-08-16T17:25:15.275Z 2018-08-16T17:25:15.275Z
#> 3             cm both    NA 2018-08-16T17:25:15.403Z 2018-08-16T17:25:15.403Z
#> 4             cm both    NA 2018-08-16T17:25:15.509Z 2018-08-16T17:25:15.509Z
#> 5             cm both    NA 2018-08-16T17:25:15.604Z 2018-08-16T17:25:15.604Z
#> 6             cm both    NA 2018-08-16T17:25:15.702Z 2018-08-16T17:25:15.702Z
#> 7             cm both    NA 2018-08-16T17:25:15.821Z 2018-08-16T17:25:15.821Z
#> 8             cm both    NA 2018-08-16T17:25:15.927Z 2018-08-16T17:25:15.927Z
#> 9             cm both    NA 2018-08-16T17:25:16.026Z 2018-08-16T17:25:16.026Z
#> 10            cm both    NA 2018-08-16T17:25:16.112Z 2018-08-16T17:25:16.112Z
#> 11            cm both    NA 2018-08-16T17:25:16.201Z 2018-08-16T17:25:16.201Z
#> 12            cm both    NA 2018-08-16T17:25:16.287Z 2018-08-16T17:25:16.287Z
#> 13            cm both    NA 2018-08-16T17:25:16.378Z 2018-08-16T17:25:16.378Z
#> 14            cm both    NA 2018-08-16T17:25:16.470Z 2018-08-16T17:25:16.470Z
#> 15            cm both    NA 2018-08-16T17:25:16.551Z 2018-08-16T17:25:16.551Z
#> 16            cm both    NA 2018-08-16T17:25:16.670Z 2018-08-16T17:25:16.670Z
#> 17            cm both    NA 2018-08-16T17:25:16.755Z 2018-08-16T17:25:16.755Z
#> 18            cm both    NA 2018-08-16T17:25:16.840Z 2018-08-16T17:25:16.840Z
#> 19            cm both    NA 2018-08-16T17:25:16.926Z 2018-08-16T17:25:16.926Z
#> 20            cm  gps    NA 2018-08-16T17:26:31.378Z 2018-08-16T17:26:31.378Z
#> 21            cm  gps    NA 2018-08-16T17:26:31.498Z 2018-08-16T17:26:31.498Z
#> 22            cm  gps    NA 2018-08-16T17:26:31.591Z 2018-08-16T17:26:31.591Z
#> 23            cm  gps    NA 2018-08-16T17:26:31.679Z 2018-08-16T17:26:31.679Z
#> 24            cm  gps    NA 2018-08-16T17:26:31.766Z 2018-08-16T17:26:31.766Z
#> 25            cm  gps    NA 2018-08-16T17:26:31.855Z 2018-08-16T17:26:31.855Z
#> 26            cm  gps    NA 2018-08-16T17:26:31.982Z 2018-08-16T17:26:31.982Z
#> 27            cm  gps    NA 2018-08-16T17:26:32.074Z 2018-08-16T17:26:32.074Z
#> 28            cm  gps    NA 2018-08-16T17:26:32.161Z 2018-08-16T17:26:32.161Z
#> 29            cm  gps    NA 2018-08-16T17:26:32.247Z 2018-08-16T17:26:32.247Z
#> 30            cm  gps    NA 2018-08-16T17:26:32.340Z 2018-08-16T17:26:32.340Z
#> 31            cm  gps    NA 2018-08-16T17:26:32.438Z 2018-08-16T17:26:32.438Z
#> 32            cm  gps    NA 2018-08-16T17:26:32.531Z 2018-08-16T17:26:32.531Z
#> 33            cm  gps    NA 2018-08-16T17:26:32.617Z 2018-08-16T17:26:32.617Z
#> 34            cm  gps    NA 2018-08-16T17:26:32.705Z 2018-08-16T17:26:32.705Z
#> 35            cm  gps    NA 2018-08-16T17:26:32.800Z 2018-08-16T17:26:32.800Z
#> 36            cm  gps    NA 2018-08-16T17:26:32.888Z 2018-08-16T17:26:32.888Z
#> 37            cm  gps    NA 2018-08-16T17:26:32.979Z 2018-08-16T17:26:32.979Z
#> 38            cm  gps    NA 2018-08-16T17:26:33.074Z 2018-08-16T17:26:33.074Z
#> 39            cm  gps    NA 2018-08-16T17:26:33.161Z 2018-08-16T17:26:33.161Z
#> 40            cm  gps    NA 2018-08-16T17:26:33.255Z 2018-08-16T17:26:33.255Z
#> 41            cm  gps    NA 2018-08-16T17:26:33.373Z 2018-08-16T17:26:33.373Z
#> 42            cm  gps    NA 2018-08-16T17:26:33.461Z 2018-08-16T17:26:33.461Z
#> 43            cm  gps    NA 2018-08-16T17:26:33.557Z 2018-08-16T17:26:33.557Z
#> 44            cm  gps    NA 2018-08-16T17:26:33.647Z 2018-08-16T17:26:33.647Z
#> 45            cm  gps    NA 2018-08-16T17:26:33.743Z 2018-08-16T17:26:33.743Z
#> 46            cm  gps    NA 2018-08-16T17:26:33.828Z 2018-08-16T17:26:33.828Z
#> 47            cm  gps    NA 2018-08-16T17:26:33.926Z 2018-08-16T17:26:33.926Z
#> 48            cm  gps    NA 2018-08-16T17:26:34.058Z 2018-08-16T17:26:34.058Z
#> 49            cm  gps    NA 2018-08-16T17:26:34.168Z 2018-08-16T17:26:34.168Z
#> 50            cm  gps    NA 2018-08-16T17:26:34.271Z 2018-08-16T17:26:34.271Z
#> 51            cm  gps    NA 2018-08-16T17:26:34.365Z 2018-08-16T17:26:34.365Z
#> 52            cm  gps    NA 2018-08-16T17:26:34.468Z 2018-08-16T17:26:34.468Z
#> 53            cm  gps    NA 2018-08-16T17:26:34.617Z 2018-08-16T17:26:34.617Z
#> 54            cm  gps    NA 2018-08-16T17:26:34.769Z 2018-08-16T17:26:34.769Z
#> 55            cm  gps    NA 2018-08-16T17:26:34.896Z 2018-08-16T17:26:34.896Z
#> 56            cm  gps    NA 2018-08-16T17:26:35.040Z 2018-08-16T17:26:35.040Z
#> 57            cm  gps    NA 2018-08-16T17:26:35.148Z 2018-08-16T17:26:35.148Z
#> 58            cm  gps    NA 2018-08-16T17:26:35.278Z 2018-08-16T17:26:35.278Z
#> 59            cm  gps    NA 2018-08-16T17:26:35.420Z 2018-08-16T17:26:35.420Z
#> 60            cm  gps    NA 2018-08-16T17:26:35.557Z 2018-08-16T17:26:35.557Z
#> 61            cm  gps    NA 2018-08-16T17:26:35.687Z 2018-08-16T17:26:35.687Z
#>    trap_id device_id lure_id geom.type  geom.coordinates
#> 1       NA        NA      NA     Point 46.7941, -72.3030
#> 2       NA        NA      NA     Point 45.4599, -71.0001
#> 3       NA        NA      NA     Point 47.3074, -71.1623
#> 4       NA        NA      NA     Point 46.6723, -71.0298
#> 5       NA        NA      NA     Point 48.2163, -71.2666
#> 6       NA        NA      NA     Point 45.8589, -71.1880
#> 7       NA        NA      NA     Point 46.7113, -70.7035
#> 8       NA        NA      NA     Point 46.2503, -71.9497
#> 9       NA        NA      NA     Point 47.1786, -70.8145
#> 10      NA        NA      NA     Point 46.9580, -71.6948
#> 11      NA        NA      NA     Point 46.7940, -72.3031
#> 12      NA        NA      NA     Point 45.4599, -71.0000
#> 13      NA        NA      NA     Point 47.3074, -71.1623
#> 14      NA        NA      NA     Point 48.2163, -71.2666
#> 15      NA        NA      NA     Point 45.8589, -71.1881
#> 16      NA        NA      NA     Point 46.7112, -70.7033
#> 17      NA        NA      NA     Point 46.2502, -71.9497
#> 18      NA        NA      NA     Point 47.1786, -70.8146
#> 19      NA        NA      NA     Point 46.9579, -71.6948
#> 20       1        NA      NA     Point 46.7942, -72.3030
#> 21       2        NA      NA     Point 48.2164, -71.2668
#> 22       3        NA      NA     Point 46.6723, -71.0297
#> 23       4        NA      NA     Point 47.3073, -71.1623
#> 24       5        NA      NA     Point 45.4599, -71.0002
#> 25       6        NA      NA     Point 46.7110, -70.7036
#> 26       7        NA      NA     Point 47.1785, -70.8147
#> 27       8        NA      NA     Point 45.8590, -71.1879
#> 28       9        NA      NA     Point 46.9579, -71.6948
#> 29      10        NA      NA     Point 47.6665, -70.7793
#> 30      11        NA      NA     Point 62.2018, -75.6396
#> 31      12        NA      NA     Point 46.2503, -71.9498
#> 32      13        NA      NA     Point 46.7111, -70.7035
#> 33      14        NA      NA     Point 58.6461, -69.9971
#> 34      15        NA      NA     Point 46.7940, -72.3031
#> 35      16        NA      NA     Point 48.2164, -71.2666
#> 36      17        NA      NA     Point 46.7822, -71.0368
#> 37      18        NA      NA     Point 47.3073, -71.1622
#> 38      19        NA      NA     Point 45.4599, -71.0001
#> 39      20        NA      NA     Point 46.7110, -70.7034
#> 40      21        NA      NA     Point 47.1786, -70.8144
#> 41      22        NA      NA     Point 45.8590, -71.1878
#> 42      23        NA      NA     Point 46.9581, -71.6949
#> 43      24        NA      NA     Point 47.6667, -70.7793
#> 44      25        NA      NA     Point 62.2018, -75.6395
#> 45      26        NA      NA     Point 46.2502, -71.9498
#> 46      27        NA      NA     Point 46.7113, -70.7035
#> 47      28        NA      NA     Point 58.6462, -69.9970
#> 48      29        NA      NA     Point 46.7941, -72.3034
#> 49      30        NA      NA     Point 48.2163, -71.2664
#> 50      31        NA      NA     Point 46.6722, -71.0299
#> 51      32        NA      NA     Point 47.3075, -71.1624
#> 52      33        NA      NA     Point 45.4599, -70.9999
#> 53      34        NA      NA     Point 46.7114, -70.7038
#> 54      35        NA      NA     Point 47.1786, -70.8147
#> 55      36        NA      NA     Point   45.859, -71.188
#> 56      37        NA      NA     Point 46.9580, -71.6945
#> 57      38        NA      NA     Point 47.6665, -70.7791
#> 58      39        NA      NA     Point 62.2018, -75.6393
#> 59      40        NA      NA     Point 46.2504, -71.9497
#> 60      41        NA      NA     Point 46.7114, -70.7034
#> 61      42        NA      NA     Point 58.6463, -69.9970
#> 
#> $response
#> Response [http://localhost:3001/api/v1/landmarks?page=0]
#>   Date: 2018-08-16 18:40
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 21 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"