Retourne les entrées présents dans la table observation

Retourne les entrées présents dans la table observation

get_obs(site_code = NULL, opened_at = NULL, closed_at = NULL,
  type = NULL, ...)

Arguments

site_code
character vecteur contenant les codes de sites pour lesquelles ont désire récupérer les campagnes (ex. 134_111_F01)

opened_at
character vecteur contenant les dates de début de campagnes (ex. 2017-01-30)

closed_at
character vecteur contenant les dates de fin de campagnes (ex. 2017-01-30)

type
character vecteur contenant le type de campagnes d'inventaires réalisé (ex. végétation)

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

head(get_obs(type="végétation"))

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 8 x 16
#>      id cell_code site_code   opened_at  closed_at type     sample_id date_obs 
#> * <int> <chr>     <chr>       <chr>      <chr>     <chr>    <lgl>     <chr>    
#> 1     2 135_104   135_104_H01 2016-06-23 2016-06-… végétat… NA        2016-06-…
#> 2     5 135_104   135_104_H01 2016-06-23 2016-06-… végétat… NA        2016-06-…
#> 3     8 135_104   135_104_H01 2016-06-23 2016-06-… végétat… NA        2016-06-…
#> 4     6 135_104   135_104_H01 2016-06-23 2016-06-… végétat… NA        2016-06-…
#> 5     4 135_104   135_104_H01 2016-06-23 2016-06-… végétat… NA        2016-06-…
#> 6     1 135_104   135_104_H01 2016-06-23 2016-06-… végétat… NA        2016-06-…
#> 7     3 135_104   135_104_H01 2016-06-23 2016-06-… végétat… NA        2016-06-…
#> 8     7 135_104   135_104_H01 2016-06-23 2016-06-… végétat… NA        2016-06-…
#> # ... with 8 more variables: obs_species.taxa_name <chr>,
#> #   obs_species.variable <chr>, obs_species.value <dbl>, is_valid <lgl>,
#> #   media <list>, notes <lgl>, created_at <chr>, updated_at <chr>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/observations?campaign_id=1&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 3.31 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> $body
#> # A tibble: 15 x 16
#>       id cell_code site_code   opened_at  closed_at type    sample_id date_obs 
#>  * <int> <chr>     <chr>       <chr>      <chr>     <chr>   <lgl>     <chr>    
#>  1    20 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#>  2    11 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#>  3    17 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#>  4    12 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#>  5    10 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#>  6    18 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#>  7    15 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#>  8    13 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#>  9    21 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#> 10    19 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#> 11    16 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#> 12    23 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#> 13    22 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#> 14    14 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#> 15     9 135_104   135_104_F01 2016-05-20 2016-07-… végéta… NA        2016-05-…
#> # ... with 8 more variables: obs_species.taxa_name <chr>,
#> #   obs_species.variable <chr>, obs_species.value <dbl>, is_valid <lgl>,
#> #   media <list>, notes <lgl>, created_at <chr>, updated_at <chr>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/observations?campaign_id=2&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 6.28 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> $body
#> # A tibble: 11 x 16
#>       id cell_code site_code   opened_at  closed_at type    sample_id date_obs 
#>  * <int> <chr>     <chr>       <chr>      <chr>     <chr>   <lgl>     <chr>    
#>  1    25 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#>  2    26 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#>  3    27 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#>  4    33 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#>  5    31 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#>  6    34 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#>  7    32 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#>  8    24 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#>  9    28 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#> 10    30 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#> 11    29 148_101   148_101_F01 2016-05-17 2016-07-… végéta… NA        2016-05-…
#> # ... with 8 more variables: obs_species.taxa_name <chr>,
#> #   obs_species.variable <chr>, obs_species.value <dbl>, is_valid <lgl>,
#> #   media <list>, notes <lgl>, created_at <chr>, updated_at <chr>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/observations?campaign_id=3&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 4.62 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#> $body
#> # A tibble: 17 x 16
#>       id cell_code site_code   opened_at  closed_at type    sample_id date_obs 
#>  * <int> <chr>     <chr>       <chr>      <chr>     <chr>   <lgl>     <chr>    
#>  1    39 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#>  2    47 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#>  3    46 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#>  4    40 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#>  5    37 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#>  6    38 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#>  7    48 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#>  8    50 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#>  9    51 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> 10    42 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> 11    41 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> 12    36 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> 13    44 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> 14    49 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> 15    45 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> 16    43 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> 17    35 137_111   137_111_H01 2016-07-11 2016-07-… végéta… NA        2016-07-…
#> # ... with 8 more variables: obs_species.taxa_name <chr>,
#> #   obs_species.variable <chr>, obs_species.value <dbl>, is_valid <lgl>,
#> #   media <list>, notes <lgl>, created_at <chr>, updated_at <chr>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/observations?campaign_id=4&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 7.12 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> [[5]]
#> [[5]][[1]]
#> $body
#> # A tibble: 18 x 16
#>       id cell_code site_code   opened_at  closed_at type    sample_id date_obs 
#>  * <int> <chr>     <chr>       <chr>      <chr>     <chr>   <lgl>     <chr>    
#>  1    66 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#>  2    57 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#>  3    64 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#>  4    56 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#>  5    65 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#>  6    52 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#>  7    55 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#>  8    68 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#>  9    62 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 10    67 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 11    69 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 12    59 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 13    54 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 14    53 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 15    58 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 16    60 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 17    61 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> 18    63 148_101   148_101_H01 2016-06-21 2016-06-… végéta… NA        2016-06-…
#> # ... with 8 more variables: obs_species.taxa_name <chr>,
#> #   obs_species.variable <chr>, obs_species.value <dbl>, is_valid <lgl>,
#> #   media <list>, notes <lgl>, created_at <chr>, updated_at <chr>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/observations?campaign_id=5&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 7.52 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> [[6]]
#> [[6]][[1]]
#> $body
#> # A tibble: 15 x 16
#>       id cell_code site_code   opened_at  closed_at type    sample_id date_obs 
#>  * <int> <chr>     <chr>       <chr>      <chr>     <chr>   <lgl>     <chr>    
#>  1    82 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#>  2    71 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#>  3    72 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#>  4    83 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#>  5    77 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#>  6    73 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#>  7    78 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#>  8    80 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#>  9    76 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#> 10    81 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#> 11    79 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#> 12    84 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#> 13    74 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#> 14    70 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#> 15    75 137_111   137_111_F01 2016-06-14 2016-07-… végéta… NA        2016-06-…
#> # ... with 8 more variables: obs_species.taxa_name <chr>,
#> #   obs_species.variable <chr>, obs_species.value <dbl>, is_valid <lgl>,
#> #   media <list>, notes <lgl>, created_at <chr>, updated_at <chr>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/observations?campaign_id=6&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 6.29 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#>