Obtenir les informations sur les sites depuis l'API de coleo

Obtenir les informations sur les sites depuis l'API de coleo

get_sites(site_code = NULL, ...)

Arguments

site_code
character vecteur contenant les identifiants uniques que l'on désire obtenir. Si site_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_sites(site_code=c('135_104_H01'))

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 1 x 19
#>      id cell_id off_station_code… site_code  type  opened_at notes created_at  
#> * <int>   <int> <lgl>             <chr>      <chr> <chr>     <lgl> <chr>       
#> 1     1     134 NA                135_104_H… mara… 2016-06-… NA    2018-08-28T…
#> # ... with 11 more variables: updated_at <chr>, campaigns <list>,
#> #   geom.type <chr>, geom.coordinates <list>, cell.id <int>, cell.name <chr>,
#> #   cell.cell_code <chr>, cell.created_at <chr>, cell.updated_at <chr>,
#> #   cell.geom.type <chr>, cell.geom.coordinates <list>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/sites?site_code=135_104_H01&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 1.32 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"

head(get_sites())

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 28 x 19
#>       id cell_id off_station_code… site_code  type  opened_at notes created_at 
#>  * <int>   <int> <chr>             <chr>      <chr> <chr>     <lgl> <chr>      
#>  1     1     134 <NA>              135_104_H… mara… 2016-06-… NA    2018-08-28…
#>  2     2     134 46487218-2        135_104_F… fore… 2016-05-… NA    2018-08-28…
#>  3     3     198 <NA>              148_101_F… fore… 2016-05-… NA    2018-08-28…
#>  4     4     151 <NA>              137_111_H… tour… 2016-07-… NA    2018-08-28…
#>  5     5     198 <NA>              148_101_H… tour… 2016-06-… NA    2018-08-28…
#>  6     6     151 471827710948-1    137_111_F… fore… 2016-06-… NA    2018-08-28…
#>  7     7     165 46407102-1        141_108_F… fore… 2016-05-… NA    2018-08-28…
#>  8     8     165 <NA>              141_108_H… tour… 2016-07-… NA    2018-08-28…
#>  9     9     119 48137116-1        132_116_F… fore… 2016-06-… NA    2018-08-28…
#> 10    10     119 <NA>              132_116_H… tour… 2016-06-… NA    2018-08-28…
#> # ... with 18 more rows, and 11 more variables: updated_at <chr>,
#> #   campaigns <list>, geom.type <chr>, geom.coordinates <list>, cell.id <int>,
#> #   cell.name <chr>, cell.cell_code <chr>, cell.created_at <chr>,
#> #   cell.updated_at <chr>, cell.geom.type <chr>, cell.geom.coordinates <list>
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/sites?page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 30.7 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#>