Retourne les entrées d'espèces présentes dans la base de données

Retourne les entrées d'espèces présentes dans la base de données

get_species(name = NULL, rank = NULL, vernacular_fr = NULL, ...)

Arguments

name
character vecteur contenant les noms taxonomiques recherché (sensible à la case) - Obsolète si l'argument query est spécifié

rank
character vecteur contenant les rangs taxonomiques recherché (sensible à la case) - Obsolète si l'argument query est spécifié

vernacular_fr
character vecteur contenant les noms vernaculaires en francais des espèces recherchées (sensible à la case) - Obsolète si l'argument query est spécifié

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_species(vernacular_fr=c('Érable'))

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 0 x 0
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?vernacular_fr=%C3%89rable&page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 2 B
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"

head(get_species())

#> [[1]]
#> [[1]][[1]]
#> $body
#> # A tibble: 100 x 6
#>    name                  vernacular_fr   rank     tsn created_at   updated_at  
#>  * <chr>                 <chr>           <chr>  <int> <chr>        <chr>       
#>  1 Ctenucha virginica    cténuche de vi… espè… 188644 2018-08-28T… 2018-08-28T…
#>  2 Boloria chariclea     boloria pourpré espè… 188571 2018-08-28T… 2018-08-28T…
#>  3 Hemaris thysbe        sphinx colibri  espè… 188633 2018-08-28T… 2018-08-28T…
#>  4 Hemaris diffinis      sphinx chèvref… espè… 936068 2018-08-28T… 2018-08-28T…
#>  5 Amelanchier           amélanchier     genre  25108 2018-08-28T… 2018-08-28T…
#>  6 Alnus rugosa          aulne rugueux   espè… 181888 2018-08-28T… 2018-08-28T…
#>  7 Betula papyrifera     bouleau à papi… espè…  19489 2018-08-28T… 2018-08-28T…
#>  8 Betula populifolia    bouleau gris    espè…  19497 2018-08-28T… 2018-08-28T…
#>  9 Betula alleghaniensis bouleau jaune   espè…  19481 2018-08-28T… 2018-08-28T…
#> 10 Carya cordiformis     caryer cordifo… espè…  19227 2018-08-28T… 2018-08-28T…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?page=0]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 17.3 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[2]]
#> $body
#> # A tibble: 100 x 6
#>    name                        vernacular_fr rank     tsn created_at updated_at
#>  * <chr>                       <chr>         <chr>  <int> <chr>      <chr>     
#>  1 Platanthera blephariglottis platanthère … espè…  43419 2018-08-2… 2018-08-2…
#>  2 Cornus canadensis           quatre-temps  espè…  27816 2018-08-2… 2018-08-2…
#>  3 Rhododendron canadense      rhododendron… espè…  23711 2018-08-2… 2018-08-2…
#>  4 Rhynchospora alba           rhynchospore… espè…  40151 2018-08-2… 2018-08-2…
#>  5 Rubus pubescens             ronce pubesc… espè…  25028 2018-08-2… 2018-08-2…
#>  6 Sarracenia purpurea         sarracénie p… espè…  21993 2018-08-2… 2018-08-2…
#>  7 Salix pyrifolia             saule baumier espè…  22575 2018-08-2… 2018-08-2…
#>  8 Maianthemum racemosum       smilacine à … espè… 503655 2018-08-2… 2018-08-2…
#>  9 Maianthemum trifolium       smilacine tr… espè… 503657 2018-08-2… 2018-08-2…
#> 10 Spiraea alba                spirée à lar… espè…  25329 2018-08-2… 2018-08-2…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?page=1]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 16.9 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[3]]
#> $body
#> # A tibble: 100 x 6
#>    name                  vernacular_fr   rank     tsn created_at   updated_at  
#>  * <chr>                 <chr>           <chr>  <int> <chr>        <chr>       
#>  1 Erebia anyuica        <NA>            espè… 778138 2018-08-28T… 2018-08-28T…
#>  2 Erebia disa           <NA>            espè… 188563 2018-08-28T… 2018-08-28T…
#>  3 Erebia discoidalis    alpin à disque… espè… 778134 2018-08-28T… 2018-08-28T…
#>  4 Erebia epipsodea      alpin commun    espè… 778133 2018-08-28T… 2018-08-28T…
#>  5 Erebia fasciata       <NA>            espè… 778132 2018-08-28T… 2018-08-28T…
#>  6 Erebia lafontainei    <NA>            espè… 778137 2018-08-28T… 2018-08-28T…
#>  7 Erebia mackinleyensis <NA>            espè… 778131 2018-08-28T… 2018-08-28T…
#>  8 Erebia magdalena      <NA>            espè… 778130 2018-08-28T… 2018-08-28T…
#>  9 Erebia mancinus       alpin à ocelle… espè… 778129 2018-08-28T… 2018-08-28T…
#> 10 Erebia pawlowskii     <NA>            espè… 778148 2018-08-28T… 2018-08-28T…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?page=2]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 16.5 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[4]]
#> $body
#> # A tibble: 100 x 6
#>    name                 vernacular_fr    rank     tsn created_at   updated_at  
#>  * <chr>                <chr>            <chr>  <int> <chr>        <chr>       
#>  1 Nymphalis vaualbum   grande vanesse   espè… 188593 2018-08-28T… 2018-08-28T…
#>  2 Oarisma garita       <NA>             espè… 706687 2018-08-28T… 2018-08-28T…
#>  3 Oarisma poweshiek    <NA>             espè… 706688 2018-08-28T… 2018-08-28T…
#>  4 Ochlodes sylvanoides <NA>             espè… 706633 2018-08-28T… 2018-08-28T…
#>  5 Oeneis alberta       <NA>             espè… 778144 2018-08-28T… 2018-08-28T…
#>  6 Oeneis alpina        <NA>             espè… 778142 2018-08-28T… 2018-08-28T…
#>  7 Oeneis bore          nordique à nerv… espè… 778143 2018-08-28T… 2018-08-28T…
#>  8 Oeneis chryxus       nordique orangé  espè… 188555 2018-08-28T… 2018-08-28T…
#>  9 Oeneis jutta         nordique des to… espè… 188557 2018-08-28T… 2018-08-28T…
#> 10 Oeneis macounii      nordique de mac… espè… 778146 2018-08-28T… 2018-08-28T…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?page=3]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 16.7 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[5]]
#> $body
#> # A tibble: 100 x 6
#>    name                 vernacular_fr   rank      tsn created_at   updated_at  
#>  * <chr>                <chr>           <chr>   <int> <chr>        <chr>       
#>  1 Speyeria zerene      <NA>            espèce 201285 2018-08-28T… 2018-08-28T…
#>  2 Staphylus hayhurstii <NA>            espèce 706785 2018-08-28T… 2018-08-28T…
#>  3 Strymon melinus      porte-queue gr… espèce 117263 2018-08-28T… 2018-08-28T…
#>  4 Thorybes bathyllus   <NA>            espèce 706829 2018-08-28T… 2018-08-28T…
#>  5 Thorybes pylades     hespérie nuage… espèce 706834 2018-08-28T… 2018-08-28T…
#>  6 Thymelicus lineola   hespérie des g… espèce 188609 2018-08-28T… 2018-08-28T…
#>  7 Urbanus proteus      <NA>            espèce 117290 2018-08-28T… 2018-08-28T…
#>  8 Vacciniina optilete  <NA>            espèce 778259 2018-08-28T… 2018-08-28T…
#>  9 Vanessa annabella    <NA>            espèce 778036 2018-08-28T… 2018-08-28T…
#> 10 Vanessa atalanta     vulcain         espèce 188599 2018-08-28T… 2018-08-28T…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?page=4]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 17.4 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[6]]
#> $body
#> # A tibble: 100 x 6
#>    name                        vernacular_fr rank     tsn created_at updated_at
#>  * <chr>                       <chr>         <chr>  <int> <chr>      <chr>     
#>  1 Carex michauxiana           carex de mic… espè…  39695 2018-08-2… 2018-08-2…
#>  2 Solidago                    verge d'or s… genre  36223 2018-08-2… 2018-08-2…
#>  3 Tussilago farfara           tussilage pa… espè…  38583 2018-08-2… 2018-08-2…
#>  4 Parthenocissus quinquefolia vigne vierge… espè…  28602 2018-08-2… 2018-08-2…
#>  5 Coleoptera                  coléoptère    ordre 109216 2018-08-2… 2018-08-2…
#>  6 Carabidae                   carabes       fami… 109234 2018-08-2… 2018-08-2…
#>  7 Arachnida                   araignée      clas…  82708 2018-08-2… 2018-08-2…
#>  8 Formicidae                  fourmis       fami… 154193 2018-08-2… 2018-08-2…
#>  9 "Orthoptera  "              orthoptère    ordre 102160 2018-08-2… 2018-08-2…
#> 10 Collembola                  collembole    ordre  99237 2018-08-2… 2018-08-2…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?page=5]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 16.9 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[7]]
#> $body
#> # A tibble: 100 x 6
#>    name                    vernacular_fr  rank      tsn created_at  updated_at 
#>  * <chr>                   <chr>          <chr>   <int> <chr>       <chr>      
#>  1 Anax longipes           anax ardent    espèce 101599 2018-08-28… 2018-08-28…
#>  2 Basiaeschna janata      aeschne print… espèce 101649 2018-08-28… 2018-08-28…
#>  3 Boyeria grafiana        aeschne fulig… espèce 101646 2018-08-28… 2018-08-28…
#>  4 Boyeria vinosa          aeschne vineu… espèce 101647 2018-08-28… 2018-08-28…
#>  5 Epiaeschna heros        aeschne majes… espèce 101638 2018-08-28… 2018-08-28…
#>  6 Gomphaeschna furcillata aeschne pygmée espèce 101635 2018-08-28… 2018-08-28…
#>  7 Nasiaeschna pentacantha aeschne cyrano espèce 101654 2018-08-28… 2018-08-28…
#>  8 Rhionaeschna mutata     aeschne des n… espèce 722026 2018-08-28… 2018-08-28…
#>  9 Gomphidae               gomphides      famil… 101664 2018-08-28… 2018-08-28…
#> 10 Arigomphus cornutus     gomphe cornu   espèce 101772 2018-08-28… 2018-08-28…
#> # ... with 90 more rows
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?page=6]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 17.7 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> [[1]][[8]]
#> $body
#> # A tibble: 3 x 6
#>   name                  vernacular_fr        rank     tsn created_at updated_at
#> * <chr>                 <chr>                <chr>  <int> <chr>      <chr>     
#> 1 Sympetrum semicinctum sympétrum semi-ambré espè… 101990 2018-08-2… 2018-08-2…
#> 2 Sympetrum vicinum     sympétrum tardif     espè… 101979 2018-08-2… 2018-08-2…
#> 3 Tramea lacerata       traméa lacérée       espè… 101822 2018-08-2… 2018-08-2…
#> 
#> $response
#> Response [https://coleo.biodiversite-quebec.ca/api/v1/taxa?page=7]
#>   Date: 2018-09-26 20:58
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 533 B
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#>