Retourne les entrées pour les pièges présents dans la base de données

Retourne les entrées pour les pièges présents dans la base de données

get_traps(trap_code = NULL, campaign_id = NULL, ...)

Arguments

trap_code
character vecteur contenant les codes de trappes désirés

campaign_id
integer vecteur contenant les identifiants uniques de la campagne

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_traps(trap_code=c('MP-22'))

#> [[1]]
#> [[1]][[1]]
#> $body
#>   id trap_code campaign_id notes               created_at
#> 1 11     MP-22          51    NA 2018-08-15T19:52:32.596Z
#>                 updated_at
#> 1 2018-08-15T19:52:32.596Z
#>                                                                                                                            landmarks
#> 1 30, 51, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.600Z, 2018-08-15T19:52:32.600Z, 11, NA, NA, Point, 62.2018, -75.6396
#>                                                                                                                                                                       samples
#> 1 6, 48, 2017-0037, 2017-0040, 2017-07-28, 2017-08-21, 11, 11, NA, NA, 2018-08-15T19:52:36.196Z, 2018-08-15T19:52:41.128Z, 2018-08-15T19:52:36.196Z, 2018-08-15T19:52:41.128Z
#> 
#> $response
#> Response [http://localhost:3001/api/v1/traps?trap_code=MP-22&page=0]
#>   Date: 2018-08-15 21:57
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 834 B
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"

get_traps()

#> [[1]]
#> [[1]][[1]]
#> $body
#>    id trap_code campaign_id notes               created_at
#> 1   1     MP-05          41    NA 2018-08-15T19:52:31.668Z
#> 2   2     MP-13          42    NA 2018-08-15T19:52:31.756Z
#> 3   3     MP-10          43    NA 2018-08-15T19:52:31.847Z
#> 4   4     MP-03          44    NA 2018-08-15T19:52:31.934Z
#> 5   5     MP-04          45    NA 2018-08-15T19:52:32.024Z
#> 6   6    MPL-01          46    NA 2018-08-15T19:52:32.115Z
#> 7   7     MP-13          47    NA 2018-08-15T19:52:32.202Z
#> 8   8     MP-01          48    NA 2018-08-15T19:52:32.288Z
#> 9   9     MP-06          49    NA 2018-08-15T19:52:32.375Z
#> 10 10     MP-21          50    NA 2018-08-15T19:52:32.502Z
#> 11 11     MP-22          51    NA 2018-08-15T19:52:32.596Z
#> 12 12     MP-11          52    NA 2018-08-15T19:52:32.684Z
#> 13 13     MP-10          53    NA 2018-08-15T19:52:32.790Z
#> 14 14     MP-16          54    NA 2018-08-15T19:52:32.878Z
#> 15 15     MP-06          41    NA 2018-08-15T19:52:32.970Z
#> 16 16     MP-12          42    NA 2018-08-15T19:52:33.057Z
#> 17 17     MP-09          43    NA 2018-08-15T19:52:33.144Z
#> 18 18     MP-11          44    NA 2018-08-15T19:52:33.234Z
#> 19 19     MP-01          45    NA 2018-08-15T19:52:33.331Z
#> 20 20    MPL-03          46    NA 2018-08-15T19:52:33.426Z
#> 21 21     MP-09          47    NA 2018-08-15T19:52:33.512Z
#> 22 22     MP-02          48    NA 2018-08-15T19:52:33.600Z
#> 23 23     MP-05          49    NA 2018-08-15T19:52:33.685Z
#> 24 24     MP-20          50    NA 2018-08-15T19:52:33.805Z
#> 25 25     MP-29          51    NA 2018-08-15T19:52:33.895Z
#> 26 26     MP-04          52    NA 2018-08-15T19:52:33.984Z
#> 27 27     MP-08          53    NA 2018-08-15T19:52:34.073Z
#> 28 28     MP-18          54    NA 2018-08-15T19:52:34.169Z
#> 29 29     MP-07          41    NA 2018-08-15T19:52:34.257Z
#> 30 30     MP-14          42    NA 2018-08-15T19:52:34.347Z
#> 31 31     MP-08          43    NA 2018-08-15T19:52:34.438Z
#> 32 32     MP-15          44    NA 2018-08-15T19:52:34.530Z
#> 33 33     MP-02          45    NA 2018-08-15T19:52:34.618Z
#> 34 34    MPL-02          46    NA 2018-08-15T19:52:34.706Z
#> 35 35     MP-12          47    NA 2018-08-15T19:52:34.791Z
#> 36 36     MP-03          48    NA 2018-08-15T19:52:34.881Z
#> 37 37     MP-15          49    NA 2018-08-15T19:52:34.973Z
#> 38 38     MP-17          50    NA 2018-08-15T19:52:35.090Z
#> 39 39     MP-23          51    NA 2018-08-15T19:52:35.178Z
#> 40 40     MP-14          52    NA 2018-08-15T19:52:35.270Z
#> 41 41     MP-07          53    NA 2018-08-15T19:52:35.368Z
#> 42 42     MP-19          54    NA 2018-08-15T19:52:35.459Z
#>                  updated_at
#> 1  2018-08-15T19:52:31.668Z
#> 2  2018-08-15T19:52:31.756Z
#> 3  2018-08-15T19:52:31.847Z
#> 4  2018-08-15T19:52:31.934Z
#> 5  2018-08-15T19:52:32.024Z
#> 6  2018-08-15T19:52:32.115Z
#> 7  2018-08-15T19:52:32.202Z
#> 8  2018-08-15T19:52:32.288Z
#> 9  2018-08-15T19:52:32.375Z
#> 10 2018-08-15T19:52:32.502Z
#> 11 2018-08-15T19:52:32.596Z
#> 12 2018-08-15T19:52:32.684Z
#> 13 2018-08-15T19:52:32.790Z
#> 14 2018-08-15T19:52:32.878Z
#> 15 2018-08-15T19:52:32.970Z
#> 16 2018-08-15T19:52:33.057Z
#> 17 2018-08-15T19:52:33.144Z
#> 18 2018-08-15T19:52:33.234Z
#> 19 2018-08-15T19:52:33.331Z
#> 20 2018-08-15T19:52:33.426Z
#> 21 2018-08-15T19:52:33.512Z
#> 22 2018-08-15T19:52:33.600Z
#> 23 2018-08-15T19:52:33.685Z
#> 24 2018-08-15T19:52:33.805Z
#> 25 2018-08-15T19:52:33.895Z
#> 26 2018-08-15T19:52:33.984Z
#> 27 2018-08-15T19:52:34.073Z
#> 28 2018-08-15T19:52:34.169Z
#> 29 2018-08-15T19:52:34.257Z
#> 30 2018-08-15T19:52:34.347Z
#> 31 2018-08-15T19:52:34.438Z
#> 32 2018-08-15T19:52:34.530Z
#> 33 2018-08-15T19:52:34.618Z
#> 34 2018-08-15T19:52:34.706Z
#> 35 2018-08-15T19:52:34.791Z
#> 36 2018-08-15T19:52:34.881Z
#> 37 2018-08-15T19:52:34.973Z
#> 38 2018-08-15T19:52:35.090Z
#> 39 2018-08-15T19:52:35.178Z
#> 40 2018-08-15T19:52:35.270Z
#> 41 2018-08-15T19:52:35.368Z
#> 42 2018-08-15T19:52:35.459Z
#>                                                                                                                             landmarks
#> 1    20, 41, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:31.672Z, 2018-08-15T19:52:31.672Z, 1, NA, NA, Point, 46.7942, -72.303
#> 2   21, 42, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:31.761Z, 2018-08-15T19:52:31.761Z, 2, NA, NA, Point, 48.2164, -71.2668
#> 3   22, 43, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:31.851Z, 2018-08-15T19:52:31.851Z, 3, NA, NA, Point, 46.6723, -71.0297
#> 4   23, 44, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:31.937Z, 2018-08-15T19:52:31.937Z, 4, NA, NA, Point, 47.3073, -71.1623
#> 5   24, 45, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.028Z, 2018-08-15T19:52:32.028Z, 5, NA, NA, Point, 45.4599, -71.0002
#> 6    25, 46, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.119Z, 2018-08-15T19:52:32.119Z, 6, NA, NA, Point, 46.711, -70.7036
#> 7   26, 47, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.206Z, 2018-08-15T19:52:32.206Z, 7, NA, NA, Point, 47.1785, -70.8147
#> 8    27, 48, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.292Z, 2018-08-15T19:52:32.292Z, 8, NA, NA, Point, 45.859, -71.1879
#> 9   28, 49, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.379Z, 2018-08-15T19:52:32.379Z, 9, NA, NA, Point, 46.9579, -71.6948
#> 10 29, 50, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.506Z, 2018-08-15T19:52:32.506Z, 10, NA, NA, Point, 47.6665, -70.7793
#> 11 30, 51, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.600Z, 2018-08-15T19:52:32.600Z, 11, NA, NA, Point, 62.2018, -75.6396
#> 12 31, 52, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.689Z, 2018-08-15T19:52:32.689Z, 12, NA, NA, Point, 46.2503, -71.9498
#> 13 32, 53, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.794Z, 2018-08-15T19:52:32.794Z, 13, NA, NA, Point, 46.7111, -70.7035
#> 14 33, 54, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.885Z, 2018-08-15T19:52:32.885Z, 14, NA, NA, Point, 58.6461, -69.9971
#> 15  34, 41, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:32.974Z, 2018-08-15T19:52:32.974Z, 15, NA, NA, Point, 46.794, -72.3031
#> 16 35, 42, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.061Z, 2018-08-15T19:52:33.061Z, 16, NA, NA, Point, 48.2164, -71.2666
#> 17 36, 43, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.148Z, 2018-08-15T19:52:33.148Z, 17, NA, NA, Point, 46.7822, -71.0368
#> 18 37, 44, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.239Z, 2018-08-15T19:52:33.239Z, 18, NA, NA, Point, 47.3073, -71.1622
#> 19 38, 45, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.335Z, 2018-08-15T19:52:33.335Z, 19, NA, NA, Point, 45.4599, -71.0001
#> 20  39, 46, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.430Z, 2018-08-15T19:52:33.430Z, 20, NA, NA, Point, 46.711, -70.7034
#> 21 40, 47, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.516Z, 2018-08-15T19:52:33.516Z, 21, NA, NA, Point, 47.1786, -70.8144
#> 22  41, 48, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.604Z, 2018-08-15T19:52:33.604Z, 22, NA, NA, Point, 45.859, -71.1878
#> 23 42, 49, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.690Z, 2018-08-15T19:52:33.690Z, 23, NA, NA, Point, 46.9581, -71.6949
#> 24 43, 50, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.809Z, 2018-08-15T19:52:33.809Z, 24, NA, NA, Point, 47.6667, -70.7793
#> 25 44, 51, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.900Z, 2018-08-15T19:52:33.900Z, 25, NA, NA, Point, 62.2018, -75.6395
#> 26 45, 52, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:33.988Z, 2018-08-15T19:52:33.988Z, 26, NA, NA, Point, 46.2502, -71.9498
#> 27 46, 53, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.077Z, 2018-08-15T19:52:34.077Z, 27, NA, NA, Point, 46.7113, -70.7035
#> 28  47, 54, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.174Z, 2018-08-15T19:52:34.174Z, 28, NA, NA, Point, 58.6462, -69.997
#> 29 48, 41, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.261Z, 2018-08-15T19:52:34.261Z, 29, NA, NA, Point, 46.7941, -72.3034
#> 30 49, 42, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.351Z, 2018-08-15T19:52:34.351Z, 30, NA, NA, Point, 48.2163, -71.2664
#> 31 50, 43, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.441Z, 2018-08-15T19:52:34.441Z, 31, NA, NA, Point, 46.6722, -71.0299
#> 32 51, 44, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.535Z, 2018-08-15T19:52:34.535Z, 32, NA, NA, Point, 47.3075, -71.1624
#> 33 52, 45, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.622Z, 2018-08-15T19:52:34.622Z, 33, NA, NA, Point, 45.4599, -70.9999
#> 34 53, 46, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.711Z, 2018-08-15T19:52:34.711Z, 34, NA, NA, Point, 46.7114, -70.7038
#> 35 54, 47, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.795Z, 2018-08-15T19:52:34.795Z, 35, NA, NA, Point, 47.1786, -70.8147
#> 36   55, 48, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.885Z, 2018-08-15T19:52:34.885Z, 36, NA, NA, Point, 45.859, -71.188
#> 37  56, 49, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:34.978Z, 2018-08-15T19:52:34.978Z, 37, NA, NA, Point, 46.958, -71.6945
#> 38 57, 50, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:35.096Z, 2018-08-15T19:52:35.096Z, 38, NA, NA, Point, 47.6665, -70.7791
#> 39 58, 51, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:35.183Z, 2018-08-15T19:52:35.183Z, 39, NA, NA, Point, 62.2018, -75.6393
#> 40 59, 52, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:35.274Z, 2018-08-15T19:52:35.274Z, 40, NA, NA, Point, 46.2504, -71.9497
#> 41 60, 53, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:35.373Z, 2018-08-15T19:52:35.373Z, 41, NA, NA, Point, 46.7114, -70.7034
#> 42  61, 54, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-15T19:52:35.464Z, 2018-08-15T19:52:35.464Z, 42, NA, NA, Point, 58.6463, -69.997
#>                                                                                                                                                                         samples
#> 1     43, 1, 2016-0010, 2016-0007, 2016-07-12, 2016-06-10, 1, 1, NA, NA, 2018-08-15T19:52:40.560Z, 2018-08-15T19:52:35.626Z, 2018-08-15T19:52:40.560Z, 2018-08-15T19:52:35.626Z
#> 2     44, 2, 2016-0016, 2016-0013, 2016-07-25, 2016-06-28, 2, 2, NA, NA, 2018-08-15T19:52:40.674Z, 2018-08-15T19:52:35.739Z, 2018-08-15T19:52:40.674Z, 2018-08-15T19:52:35.739Z
#> 3     45, 3, 2016-0022, 2016-0019, 2016-07-20, 2016-06-20, 3, 3, NA, NA, 2018-08-15T19:52:40.789Z, 2018-08-15T19:52:35.863Z, 2018-08-15T19:52:40.789Z, 2018-08-15T19:52:35.863Z
#> 4     46, 4, 2016-0028, 2016-0025, 2016-07-28, 2016-07-11, 4, 4, NA, NA, 2018-08-15T19:52:40.906Z, 2018-08-15T19:52:35.973Z, 2018-08-15T19:52:40.906Z, 2018-08-15T19:52:35.973Z
#> 5     47, 5, 2016-0004, 2016-0001, 2016-07-21, 2016-06-21, 5, 5, NA, NA, 2018-08-15T19:52:41.016Z, 2018-08-15T19:52:36.086Z, 2018-08-15T19:52:41.016Z, 2018-08-15T19:52:36.086Z
#> 6                                                                                           7, 2017-0043, 2017-07-11, 6, NA, 2018-08-15T19:52:36.313Z, 2018-08-15T19:52:36.313Z
#> 7     50, 9, 2017-0022, 2017-0019, 2017-07-10, 2017-06-20, 7, 7, NA, NA, 2018-08-15T19:52:41.359Z, 2018-08-15T19:52:36.570Z, 2018-08-15T19:52:41.359Z, 2018-08-15T19:52:36.570Z
#> 8    52, 11, 2017-0004, 2017-0001, 2017-06-29, 2017-06-08, 8, 8, NA, NA, 2018-08-15T19:52:41.619Z, 2018-08-15T19:52:36.801Z, 2018-08-15T19:52:41.619Z, 2018-08-15T19:52:36.801Z
#> 9    53, 12, 2017-0010, 2017-0007, 2017-07-10, 2017-06-20, 9, 9, NA, NA, 2018-08-15T19:52:41.728Z, 2018-08-15T19:52:36.910Z, 2018-08-15T19:52:41.728Z, 2018-08-15T19:52:36.910Z
#> 10                 54, 13, 2017-0034, 2017-0031, NA, NA, 10, 10, NA, NA, 2018-08-15T19:52:41.841Z, 2018-08-15T19:52:37.053Z, 2018-08-15T19:52:41.841Z, 2018-08-15T19:52:37.053Z
#> 11  48, 6, 2017-0040, 2017-0037, 2017-08-21, 2017-07-28, 11, 11, NA, NA, 2018-08-15T19:52:41.128Z, 2018-08-15T19:52:36.196Z, 2018-08-15T19:52:41.128Z, 2018-08-15T19:52:36.196Z
#> 12  49, 8, 2017-0028, 2017-0025, 2017-07-11, 2017-06-15, 12, 12, NA, NA, 2018-08-15T19:52:41.248Z, 2018-08-15T19:52:36.457Z, 2018-08-15T19:52:41.248Z, 2018-08-15T19:52:36.457Z
#> 13 51, 10, 2017-0016, 2017-0013, 2017-07-11, 2017-06-14, 13, 13, NA, NA, 2018-08-15T19:52:41.474Z, 2018-08-15T19:52:36.689Z, 2018-08-15T19:52:41.474Z, 2018-08-15T19:52:36.689Z
#> 14 55, 14, 2017-0049, 2017-0046, 2017-09-15, 2017-09-02, 14, 14, NA, NA, 2018-08-15T19:52:41.958Z, 2018-08-15T19:52:37.160Z, 2018-08-15T19:52:41.958Z, 2018-08-15T19:52:37.160Z
#> 15 56, 15, 2016-0011, 2016-0008, 2016-07-12, 2016-06-10, 15, 15, NA, NA, 2018-08-15T19:52:42.067Z, 2018-08-15T19:52:37.271Z, 2018-08-15T19:52:42.067Z, 2018-08-15T19:52:37.271Z
#> 16 57, 16, 2016-0017, 2016-0014, 2016-07-25, 2016-06-28, 16, 16, NA, NA, 2018-08-15T19:52:42.178Z, 2018-08-15T19:52:37.385Z, 2018-08-15T19:52:42.178Z, 2018-08-15T19:52:37.385Z
#> 17 58, 17, 2016-0023, 2016-0020, 2016-07-20, 2016-06-20, 17, 17, NA, NA, 2018-08-15T19:52:42.294Z, 2018-08-15T19:52:37.499Z, 2018-08-15T19:52:42.294Z, 2018-08-15T19:52:37.499Z
#> 18 59, 18, 2016-0029, 2016-0026, 2016-07-28, 2016-07-11, 18, 18, NA, NA, 2018-08-15T19:52:42.407Z, 2018-08-15T19:52:37.612Z, 2018-08-15T19:52:42.407Z, 2018-08-15T19:52:37.612Z
#> 19 60, 19, 2016-0005, 2016-0002, 2016-07-21, 2016-06-21, 19, 19, NA, NA, 2018-08-15T19:52:42.517Z, 2018-08-15T19:52:37.758Z, 2018-08-15T19:52:42.517Z, 2018-08-15T19:52:37.758Z
#> 20                                                                                        21, 2017-0044, 2017-07-11, 20, NA, 2018-08-15T19:52:37.986Z, 2018-08-15T19:52:37.986Z
#> 21 63, 23, 2017-0023, 2017-0020, 2017-07-10, 2017-06-20, 21, 21, NA, NA, 2018-08-15T19:52:42.876Z, 2018-08-15T19:52:38.225Z, 2018-08-15T19:52:42.876Z, 2018-08-15T19:52:38.225Z
#> 22 65, 25, 2017-0005, 2017-0002, 2017-06-29, 2017-06-08, 22, 22, NA, NA, 2018-08-15T19:52:43.107Z, 2018-08-15T19:52:38.446Z, 2018-08-15T19:52:43.107Z, 2018-08-15T19:52:38.446Z
#> 23 66, 26, 2017-0011, 2017-0008, 2017-07-10, 2017-06-20, 23, 23, NA, NA, 2018-08-15T19:52:43.214Z, 2018-08-15T19:52:38.566Z, 2018-08-15T19:52:43.214Z, 2018-08-15T19:52:38.566Z
#> 24                 67, 27, 2017-0035, 2017-0032, NA, NA, 24, 24, NA, NA, 2018-08-15T19:52:43.323Z, 2018-08-15T19:52:38.682Z, 2018-08-15T19:52:43.323Z, 2018-08-15T19:52:38.682Z
#> 25 61, 20, 2017-0041, 2017-0038, 2017-08-21, 2017-07-28, 25, 25, NA, NA, 2018-08-15T19:52:42.630Z, 2018-08-15T19:52:37.874Z, 2018-08-15T19:52:42.630Z, 2018-08-15T19:52:37.874Z
#> 26 62, 22, 2017-0029, 2017-0026, 2017-07-11, 2017-06-15, 26, 26, NA, NA, 2018-08-15T19:52:42.741Z, 2018-08-15T19:52:38.099Z, 2018-08-15T19:52:42.741Z, 2018-08-15T19:52:38.099Z
#> 27 64, 24, 2017-0017, 2017-0014, 2017-07-11, 2017-06-14, 27, 27, NA, NA, 2018-08-15T19:52:42.988Z, 2018-08-15T19:52:38.336Z, 2018-08-15T19:52:42.988Z, 2018-08-15T19:52:38.336Z
#> 28 68, 28, 2017-0050, 2017-0047, 2017-09-15, 2017-09-02, 28, 28, NA, NA, 2018-08-15T19:52:43.441Z, 2018-08-15T19:52:38.791Z, 2018-08-15T19:52:43.441Z, 2018-08-15T19:52:38.791Z
#> 29 69, 29, 2016-0012, 2016-0009, 2016-07-12, 2016-06-10, 29, 29, NA, NA, 2018-08-15T19:52:43.547Z, 2018-08-15T19:52:38.911Z, 2018-08-15T19:52:43.547Z, 2018-08-15T19:52:38.911Z
#> 30 70, 30, 2016-0018, 2016-0015, 2016-07-25, 2016-06-28, 30, 30, NA, NA, 2018-08-15T19:52:43.661Z, 2018-08-15T19:52:39.050Z, 2018-08-15T19:52:43.661Z, 2018-08-15T19:52:39.050Z
#> 31 71, 31, 2016-0024, 2016-0021, 2016-07-20, 2016-06-20, 31, 31, NA, NA, 2018-08-15T19:52:43.778Z, 2018-08-15T19:52:39.165Z, 2018-08-15T19:52:43.778Z, 2018-08-15T19:52:39.165Z
#> 32 72, 32, 2016-0030, 2016-0027, 2016-07-28, 2016-07-11, 32, 32, NA, NA, 2018-08-15T19:52:43.890Z, 2018-08-15T19:52:39.282Z, 2018-08-15T19:52:43.890Z, 2018-08-15T19:52:39.282Z
#> 33 73, 33, 2016-0006, 2016-0003, 2016-07-21, 2016-06-21, 33, 33, NA, NA, 2018-08-15T19:52:43.999Z, 2018-08-15T19:52:39.394Z, 2018-08-15T19:52:43.999Z, 2018-08-15T19:52:39.394Z
#> 34                                                                                        35, 2017-0045, 2017-07-11, 34, NA, 2018-08-15T19:52:39.615Z, 2018-08-15T19:52:39.615Z
#> 35 76, 37, 2017-0024, 2017-0021, 2017-07-10, 2017-06-20, 35, 35, NA, NA, 2018-08-15T19:52:44.368Z, 2018-08-15T19:52:39.844Z, 2018-08-15T19:52:44.368Z, 2018-08-15T19:52:39.844Z
#> 36 78, 39, 2017-0006, 2017-0003, 2017-06-29, 2017-06-08, 36, 36, NA, NA, 2018-08-15T19:52:44.592Z, 2018-08-15T19:52:40.079Z, 2018-08-15T19:52:44.592Z, 2018-08-15T19:52:40.079Z
#> 37 79, 40, 2017-0012, 2017-0009, 2017-07-10, 2017-06-20, 37, 37, NA, NA, 2018-08-15T19:52:44.702Z, 2018-08-15T19:52:40.190Z, 2018-08-15T19:52:44.702Z, 2018-08-15T19:52:40.190Z
#> 38                 80, 41, 2017-0036, 2017-0033, NA, NA, 38, 38, NA, NA, 2018-08-15T19:52:44.812Z, 2018-08-15T19:52:40.330Z, 2018-08-15T19:52:44.812Z, 2018-08-15T19:52:40.330Z
#> 39 74, 34, 2017-0042, 2017-0039, 2017-08-21, 2017-07-28, 39, 39, NA, NA, 2018-08-15T19:52:44.115Z, 2018-08-15T19:52:39.502Z, 2018-08-15T19:52:44.115Z, 2018-08-15T19:52:39.502Z
#> 40 75, 36, 2017-0030, 2017-0027, 2017-07-11, 2017-06-15, 40, 40, NA, NA, 2018-08-15T19:52:44.257Z, 2018-08-15T19:52:39.734Z, 2018-08-15T19:52:44.257Z, 2018-08-15T19:52:39.734Z
#> 41 77, 38, 2017-0018, 2017-0015, 2017-07-11, 2017-06-14, 41, 41, NA, NA, 2018-08-15T19:52:44.482Z, 2018-08-15T19:52:39.960Z, 2018-08-15T19:52:44.482Z, 2018-08-15T19:52:39.960Z
#> 42 81, 42, 2017-0051, 2017-0048, 2017-09-15, 2017-09-02, 42, 42, NA, NA, 2018-08-15T19:52:44.941Z, 2018-08-15T19:52:40.450Z, 2018-08-15T19:52:44.941Z, 2018-08-15T19:52:40.450Z
#> 
#> $response
#> Response [http://localhost:3001/api/v1/traps?page=0]
#>   Date: 2018-08-15 21:57
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 34.4 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"