Retourne les entrées pour les repères sur le terrain spécifique à une campagne d'inventaires

Retourne les entrées pour les repères sur le terrain spécifique à une campagne d'inventaires

get_landmarks(site_code = NULL, opened_at = NULL, closed_at = NULL,
  type = NULL, ...)

Arguments

site_code
character vecteur contenant les codes de sites pour lesquelles ont désire récupérer les campagnes (ex. 134_111_F01)

opened_at
character vecteur contenant les dates de début de campagnes (ex. 2017-01-30)

closed_at
character vecteur contenant les dates de fin de campagnes (ex. 2017-01-30)

type
character vecteur contenant le type de campagnes d'inventaires réalisé (ex. végétation)

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_landmarks()

#> [[1]]
#> [[1]][[1]]
#> $body
#>    id campaign_id tree_code sp_id dbh dbh_unit azimut distance distance_unit
#> 1   1           2         5    23 380       mm    200       90            cm
#> 2   2           3         1    23 350       mm     21      213            cm
#> 3   3           6         9    47 200       mm    188      276            cm
#> 4   4           7         6    23 680       mm    221      440            cm
#> 5   5           9        57    47 330       mm    274      400            cm
#> 6   6          21         6    46 300       mm    279      216            cm
#> 7   7          22         7    26 250       mm    169      230            cm
#> 8   8          23        29     8 100       mm    169      460            cm
#> 9   9          24        29     7 140       mm    172      114            cm
#> 10 10          25        17    23 260       mm     14      335            cm
#> 11 11           2        19    23 280       mm     41      234            cm
#> 12 12           3         2    26 140       mm    246      290            cm
#> 13 13           6        10    47 170       mm    230      346            cm
#> 14 14           9        37    47 250       mm    174      472            cm
#> 15 15          21        19     7 150       mm     17      459            cm
#> 16 16          22        19    23 240       mm     25      393            cm
#> 17 17          23        10    47 240       mm     69      800            cm
#> 18 18          24        58     7 160       mm     21       93            cm
#> 19 19          25         7     9 390       mm    290      560            cm
#> 20 20          41      <NA>    NA   0       mm      0        0            cm
#> 21 21          42      <NA>    NA   0       mm      0        0            cm
#> 22 22          43      <NA>    NA   0       mm      0        0            cm
#> 23 23          44      <NA>    NA   0       mm      0        0            cm
#> 24 24          45      <NA>    NA   0       mm      0        0            cm
#> 25 25          46      <NA>    NA   0       mm      0        0            cm
#> 26 26          47      <NA>    NA   0       mm      0        0            cm
#> 27 27          48      <NA>    NA   0       mm      0        0            cm
#> 28 28          49      <NA>    NA   0       mm      0        0            cm
#> 29 29          50      <NA>    NA   0       mm      0        0            cm
#> 30 30          51      <NA>    NA   0       mm      0        0            cm
#> 31 31          52      <NA>    NA   0       mm      0        0            cm
#> 32 32          53      <NA>    NA   0       mm      0        0            cm
#> 33 33          54      <NA>    NA   0       mm      0        0            cm
#> 34 34          41      <NA>    NA   0       mm      0        0            cm
#> 35 35          42      <NA>    NA   0       mm      0        0            cm
#> 36 36          43      <NA>    NA   0       mm      0        0            cm
#> 37 37          44      <NA>    NA   0       mm      0        0            cm
#> 38 38          45      <NA>    NA   0       mm      0        0            cm
#> 39 39          46      <NA>    NA   0       mm      0        0            cm
#> 40 40          47      <NA>    NA   0       mm      0        0            cm
#> 41 41          48      <NA>    NA   0       mm      0        0            cm
#> 42 42          49      <NA>    NA   0       mm      0        0            cm
#> 43 43          50      <NA>    NA   0       mm      0        0            cm
#> 44 44          51      <NA>    NA   0       mm      0        0            cm
#> 45 45          52      <NA>    NA   0       mm      0        0            cm
#> 46 46          53      <NA>    NA   0       mm      0        0            cm
#> 47 47          54      <NA>    NA   0       mm      0        0            cm
#> 48 48          41      <NA>    NA   0       mm      0        0            cm
#> 49 49          42      <NA>    NA   0       mm      0        0            cm
#> 50 50          43      <NA>    NA   0       mm      0        0            cm
#> 51 51          44      <NA>    NA   0       mm      0        0            cm
#> 52 52          45      <NA>    NA   0       mm      0        0            cm
#> 53 53          46      <NA>    NA   0       mm      0        0            cm
#> 54 54          47      <NA>    NA   0       mm      0        0            cm
#> 55 55          48      <NA>    NA   0       mm      0        0            cm
#> 56 56          49      <NA>    NA   0       mm      0        0            cm
#> 57 57          50      <NA>    NA   0       mm      0        0            cm
#> 58 58          51      <NA>    NA   0       mm      0        0            cm
#> 59 59          52      <NA>    NA   0       mm      0        0            cm
#> 60 60          53      <NA>    NA   0       mm      0        0            cm
#> 61 61          54      <NA>    NA   0       mm      0        0            cm
#>    type notes               created_at               updated_at trap_id
#> 1  both    NA 2018-08-10T17:16:15.860Z 2018-08-10T17:16:15.860Z      NA
#> 2  both    NA 2018-08-10T17:16:15.951Z 2018-08-10T17:16:15.951Z      NA
#> 3  both    NA 2018-08-10T17:16:16.054Z 2018-08-10T17:16:16.054Z      NA
#> 4  both    NA 2018-08-10T17:16:16.169Z 2018-08-10T17:16:16.169Z      NA
#> 5  both    NA 2018-08-10T17:16:16.259Z 2018-08-10T17:16:16.259Z      NA
#> 6  both    NA 2018-08-10T17:16:16.357Z 2018-08-10T17:16:16.357Z      NA
#> 7  both    NA 2018-08-10T17:16:16.464Z 2018-08-10T17:16:16.464Z      NA
#> 8  both    NA 2018-08-10T17:16:16.600Z 2018-08-10T17:16:16.600Z      NA
#> 9  both    NA 2018-08-10T17:16:16.681Z 2018-08-10T17:16:16.681Z      NA
#> 10 both    NA 2018-08-10T17:16:16.770Z 2018-08-10T17:16:16.770Z      NA
#> 11 both    NA 2018-08-10T17:16:16.864Z 2018-08-10T17:16:16.864Z      NA
#> 12 both    NA 2018-08-10T17:16:16.956Z 2018-08-10T17:16:16.956Z      NA
#> 13 both    NA 2018-08-10T17:16:17.040Z 2018-08-10T17:16:17.040Z      NA
#> 14 both    NA 2018-08-10T17:16:17.136Z 2018-08-10T17:16:17.136Z      NA
#> 15 both    NA 2018-08-10T17:16:17.217Z 2018-08-10T17:16:17.217Z      NA
#> 16 both    NA 2018-08-10T17:16:17.298Z 2018-08-10T17:16:17.298Z      NA
#> 17 both    NA 2018-08-10T17:16:17.391Z 2018-08-10T17:16:17.391Z      NA
#> 18 both    NA 2018-08-10T17:16:17.481Z 2018-08-10T17:16:17.481Z      NA
#> 19 both    NA 2018-08-10T17:16:17.603Z 2018-08-10T17:16:17.603Z      NA
#> 20  gps    NA 2018-08-10T17:20:16.957Z 2018-08-10T17:20:16.957Z       1
#> 21  gps    NA 2018-08-10T17:20:17.079Z 2018-08-10T17:20:17.079Z       2
#> 22  gps    NA 2018-08-10T17:20:17.194Z 2018-08-10T17:20:17.194Z       3
#> 23  gps    NA 2018-08-10T17:20:17.292Z 2018-08-10T17:20:17.292Z       4
#> 24  gps    NA 2018-08-10T17:20:17.396Z 2018-08-10T17:20:17.396Z       5
#> 25  gps    NA 2018-08-10T17:20:17.495Z 2018-08-10T17:20:17.495Z       6
#> 26  gps    NA 2018-08-10T17:20:17.589Z 2018-08-10T17:20:17.589Z       7
#> 27  gps    NA 2018-08-10T17:20:17.688Z 2018-08-10T17:20:17.688Z       8
#> 28  gps    NA 2018-08-10T17:20:17.777Z 2018-08-10T17:20:17.777Z       9
#> 29  gps    NA 2018-08-10T17:20:17.865Z 2018-08-10T17:20:17.865Z      10
#> 30  gps    NA 2018-08-10T17:20:17.961Z 2018-08-10T17:20:17.961Z      11
#> 31  gps    NA 2018-08-10T17:20:18.085Z 2018-08-10T17:20:18.085Z      12
#> 32  gps    NA 2018-08-10T17:20:18.179Z 2018-08-10T17:20:18.179Z      13
#> 33  gps    NA 2018-08-10T17:20:18.275Z 2018-08-10T17:20:18.275Z      14
#> 34  gps    NA 2018-08-10T17:20:18.386Z 2018-08-10T17:20:18.386Z      15
#> 35  gps    NA 2018-08-10T17:20:18.480Z 2018-08-10T17:20:18.480Z      16
#> 36  gps    NA 2018-08-10T17:20:18.608Z 2018-08-10T17:20:18.608Z      17
#> 37  gps    NA 2018-08-10T17:20:18.711Z 2018-08-10T17:20:18.711Z      18
#> 38  gps    NA 2018-08-10T17:20:18.811Z 2018-08-10T17:20:18.811Z      19
#> 39  gps    NA 2018-08-10T17:20:18.904Z 2018-08-10T17:20:18.904Z      20
#> 40  gps    NA 2018-08-10T17:20:19.000Z 2018-08-10T17:20:19.000Z      21
#> 41  gps    NA 2018-08-10T17:20:19.101Z 2018-08-10T17:20:19.101Z      22
#> 42  gps    NA 2018-08-10T17:20:19.199Z 2018-08-10T17:20:19.199Z      23
#> 43  gps    NA 2018-08-10T17:20:19.295Z 2018-08-10T17:20:19.295Z      24
#> 44  gps    NA 2018-08-10T17:20:19.393Z 2018-08-10T17:20:19.393Z      25
#> 45  gps    NA 2018-08-10T17:20:19.534Z 2018-08-10T17:20:19.534Z      26
#> 46  gps    NA 2018-08-10T17:20:19.629Z 2018-08-10T17:20:19.629Z      27
#> 47  gps    NA 2018-08-10T17:20:19.716Z 2018-08-10T17:20:19.716Z      28
#> 48  gps    NA 2018-08-10T17:20:19.807Z 2018-08-10T17:20:19.807Z      29
#> 49  gps    NA 2018-08-10T17:20:19.912Z 2018-08-10T17:20:19.912Z      30
#> 50  gps    NA 2018-08-10T17:20:20.007Z 2018-08-10T17:20:20.007Z      31
#> 51  gps    NA 2018-08-10T17:20:20.107Z 2018-08-10T17:20:20.107Z      32
#> 52  gps    NA 2018-08-10T17:20:20.219Z 2018-08-10T17:20:20.219Z      33
#> 53  gps    NA 2018-08-10T17:20:20.327Z 2018-08-10T17:20:20.327Z      34
#> 54  gps    NA 2018-08-10T17:20:20.415Z 2018-08-10T17:20:20.415Z      35
#> 55  gps    NA 2018-08-10T17:20:20.504Z 2018-08-10T17:20:20.504Z      36
#> 56  gps    NA 2018-08-10T17:20:20.595Z 2018-08-10T17:20:20.595Z      37
#> 57  gps    NA 2018-08-10T17:20:20.703Z 2018-08-10T17:20:20.703Z      38
#> 58  gps    NA 2018-08-10T17:20:20.798Z 2018-08-10T17:20:20.798Z      39
#> 59  gps    NA 2018-08-10T17:20:20.917Z 2018-08-10T17:20:20.917Z      40
#> 60  gps    NA 2018-08-10T17:20:21.008Z 2018-08-10T17:20:21.008Z      41
#> 61  gps    NA 2018-08-10T17:20:21.117Z 2018-08-10T17:20:21.117Z      42
#>    device_id lure_id geom.type  geom.coordinates
#> 1         NA      NA     Point 46.7941, -72.3030
#> 2         NA      NA     Point 45.4599, -71.0001
#> 3         NA      NA     Point 47.3074, -71.1623
#> 4         NA      NA     Point 46.6723, -71.0298
#> 5         NA      NA     Point 48.2163, -71.2666
#> 6         NA      NA     Point 45.8589, -71.1880
#> 7         NA      NA     Point 46.7113, -70.7035
#> 8         NA      NA     Point 46.2503, -71.9497
#> 9         NA      NA     Point 47.1786, -70.8145
#> 10        NA      NA     Point 46.9580, -71.6948
#> 11        NA      NA     Point 46.7940, -72.3031
#> 12        NA      NA     Point 45.4599, -71.0000
#> 13        NA      NA     Point 47.3074, -71.1623
#> 14        NA      NA     Point 48.2163, -71.2666
#> 15        NA      NA     Point 45.8589, -71.1881
#> 16        NA      NA     Point 46.7112, -70.7033
#> 17        NA      NA     Point 46.2502, -71.9497
#> 18        NA      NA     Point 47.1786, -70.8146
#> 19        NA      NA     Point 46.9579, -71.6948
#> 20        NA      NA     Point 46.7942, -72.3030
#> 21        NA      NA     Point 48.2164, -71.2668
#> 22        NA      NA     Point 46.6723, -71.0297
#> 23        NA      NA     Point 47.3073, -71.1623
#> 24        NA      NA     Point 45.4599, -71.0002
#> 25        NA      NA     Point 46.7110, -70.7036
#> 26        NA      NA     Point 47.1785, -70.8147
#> 27        NA      NA     Point 45.8590, -71.1879
#> 28        NA      NA     Point 46.9579, -71.6948
#> 29        NA      NA     Point 47.6665, -70.7793
#> 30        NA      NA     Point 62.2018, -75.6396
#> 31        NA      NA     Point 46.2503, -71.9498
#> 32        NA      NA     Point 46.7111, -70.7035
#> 33        NA      NA     Point 58.6461, -69.9971
#> 34        NA      NA     Point 46.7940, -72.3031
#> 35        NA      NA     Point 48.2164, -71.2666
#> 36        NA      NA     Point 46.7822, -71.0368
#> 37        NA      NA     Point 47.3073, -71.1622
#> 38        NA      NA     Point 45.4599, -71.0001
#> 39        NA      NA     Point 46.7110, -70.7034
#> 40        NA      NA     Point 47.1786, -70.8144
#> 41        NA      NA     Point 45.8590, -71.1878
#> 42        NA      NA     Point 46.9581, -71.6949
#> 43        NA      NA     Point 47.6667, -70.7793
#> 44        NA      NA     Point 62.2018, -75.6395
#> 45        NA      NA     Point 46.2502, -71.9498
#> 46        NA      NA     Point 46.7113, -70.7035
#> 47        NA      NA     Point 58.6462, -69.9970
#> 48        NA      NA     Point 46.7941, -72.3034
#> 49        NA      NA     Point 48.2163, -71.2664
#> 50        NA      NA     Point 46.6722, -71.0299
#> 51        NA      NA     Point 47.3075, -71.1624
#> 52        NA      NA     Point 45.4599, -70.9999
#> 53        NA      NA     Point 46.7114, -70.7038
#> 54        NA      NA     Point 47.1786, -70.8147
#> 55        NA      NA     Point   45.859, -71.188
#> 56        NA      NA     Point 46.9580, -71.6945
#> 57        NA      NA     Point 47.6665, -70.7791
#> 58        NA      NA     Point 62.2018, -75.6393
#> 59        NA      NA     Point 46.2504, -71.9497
#> 60        NA      NA     Point 46.7114, -70.7034
#> 61        NA      NA     Point 58.6463, -69.9970
#> 
#> $response
#> Response [http://localhost:3001/api/v1/landmarks?page=0]
#>   Date: 2018-08-13 15:28
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 20.4 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#>