Retourne les entrées pour les repères sur le terrain spécifique à une campagne d'inventaires

Retourne les entrées pour les repères sur le terrain spécifique à une campagne d'inventaires

get_landmarks(site_code = NULL, opened_at = NULL, closed_at = NULL,
  type = NULL, ...)

Arguments

site_code
character vecteur contenant les codes de sites pour lesquelles ont désire récupérer les campagnes (ex. 134_111_F01)

opened_at
character vecteur contenant les dates de début de campagnes (ex. 2017-01-30)

closed_at
character vecteur contenant les dates de fin de campagnes (ex. 2017-01-30)

type
character vecteur contenant le type de campagnes d'inventaires réalisé (ex. végétation)

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

if (FALSE) {
get_landmarks()
}