Retourne les entrées pour les échantillons présents dans la base de données

Retourne les entrées pour les échantillons présents dans la base de données

get_samples(sample_code = NULL, ...)

Arguments

sample_code
character vecteur contenant les codes d'identification unique des échantillons

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_samples(trap_code=c('2016-0026'))

#> [[1]]
#> [[1]][[1]]
#> $body
#>    id sample_code  date_samp trap_id notes               created_at
#> 1   1   2016-0007 2016-06-10       1    NA 2018-08-15T19:52:35.626Z
#> 2   2   2016-0013 2016-06-28       2    NA 2018-08-15T19:52:35.739Z
#> 3   3   2016-0019 2016-06-20       3    NA 2018-08-15T19:52:35.863Z
#> 4   4   2016-0025 2016-07-11       4    NA 2018-08-15T19:52:35.973Z
#> 5   5   2016-0001 2016-06-21       5    NA 2018-08-15T19:52:36.086Z
#> 6   6   2017-0037 2017-07-28      11    NA 2018-08-15T19:52:36.196Z
#> 7   7   2017-0043 2017-07-11       6    NA 2018-08-15T19:52:36.313Z
#> 8   8   2017-0025 2017-06-15      12    NA 2018-08-15T19:52:36.457Z
#> 9   9   2017-0019 2017-06-20       7    NA 2018-08-15T19:52:36.570Z
#> 10 10   2017-0013 2017-06-14      13    NA 2018-08-15T19:52:36.689Z
#> 11 11   2017-0001 2017-06-08       8    NA 2018-08-15T19:52:36.801Z
#> 12 12   2017-0007 2017-06-20       9    NA 2018-08-15T19:52:36.910Z
#> 13 13   2017-0031       <NA>      10    NA 2018-08-15T19:52:37.053Z
#> 14 14   2017-0046 2017-09-02      14    NA 2018-08-15T19:52:37.160Z
#> 15 15   2016-0008 2016-06-10      15    NA 2018-08-15T19:52:37.271Z
#> 16 16   2016-0014 2016-06-28      16    NA 2018-08-15T19:52:37.385Z
#> 17 17   2016-0020 2016-06-20      17    NA 2018-08-15T19:52:37.499Z
#> 18 18   2016-0026 2016-07-11      18    NA 2018-08-15T19:52:37.612Z
#> 19 19   2016-0002 2016-06-21      19    NA 2018-08-15T19:52:37.758Z
#> 20 20   2017-0038 2017-07-28      25    NA 2018-08-15T19:52:37.874Z
#> 21 21   2017-0044 2017-07-11      20    NA 2018-08-15T19:52:37.986Z
#> 22 22   2017-0026 2017-06-15      26    NA 2018-08-15T19:52:38.099Z
#> 23 23   2017-0020 2017-06-20      21    NA 2018-08-15T19:52:38.225Z
#> 24 24   2017-0014 2017-06-14      27    NA 2018-08-15T19:52:38.336Z
#> 25 25   2017-0002 2017-06-08      22    NA 2018-08-15T19:52:38.446Z
#> 26 26   2017-0008 2017-06-20      23    NA 2018-08-15T19:52:38.566Z
#> 27 27   2017-0032       <NA>      24    NA 2018-08-15T19:52:38.682Z
#> 28 28   2017-0047 2017-09-02      28    NA 2018-08-15T19:52:38.791Z
#> 29 29   2016-0009 2016-06-10      29    NA 2018-08-15T19:52:38.911Z
#> 30 30   2016-0015 2016-06-28      30    NA 2018-08-15T19:52:39.050Z
#> 31 31   2016-0021 2016-06-20      31    NA 2018-08-15T19:52:39.165Z
#> 32 32   2016-0027 2016-07-11      32    NA 2018-08-15T19:52:39.282Z
#> 33 33   2016-0003 2016-06-21      33    NA 2018-08-15T19:52:39.394Z
#> 34 34   2017-0039 2017-07-28      39    NA 2018-08-15T19:52:39.502Z
#> 35 35   2017-0045 2017-07-11      34    NA 2018-08-15T19:52:39.615Z
#> 36 36   2017-0027 2017-06-15      40    NA 2018-08-15T19:52:39.734Z
#> 37 37   2017-0021 2017-06-20      35    NA 2018-08-15T19:52:39.844Z
#> 38 38   2017-0015 2017-06-14      41    NA 2018-08-15T19:52:39.960Z
#> 39 39   2017-0003 2017-06-08      36    NA 2018-08-15T19:52:40.079Z
#> 40 40   2017-0009 2017-06-20      37    NA 2018-08-15T19:52:40.190Z
#> 41 41   2017-0033       <NA>      38    NA 2018-08-15T19:52:40.330Z
#> 42 42   2017-0048 2017-09-02      42    NA 2018-08-15T19:52:40.450Z
#> 43 43   2016-0010 2016-07-12       1    NA 2018-08-15T19:52:40.560Z
#> 44 44   2016-0016 2016-07-25       2    NA 2018-08-15T19:52:40.674Z
#> 45 45   2016-0022 2016-07-20       3    NA 2018-08-15T19:52:40.789Z
#> 46 46   2016-0028 2016-07-28       4    NA 2018-08-15T19:52:40.906Z
#> 47 47   2016-0004 2016-07-21       5    NA 2018-08-15T19:52:41.016Z
#> 48 48   2017-0040 2017-08-21      11    NA 2018-08-15T19:52:41.128Z
#> 49 49   2017-0028 2017-07-11      12    NA 2018-08-15T19:52:41.248Z
#> 50 50   2017-0022 2017-07-10       7    NA 2018-08-15T19:52:41.359Z
#> 51 51   2017-0016 2017-07-11      13    NA 2018-08-15T19:52:41.474Z
#> 52 52   2017-0004 2017-06-29       8    NA 2018-08-15T19:52:41.619Z
#> 53 53   2017-0010 2017-07-10       9    NA 2018-08-15T19:52:41.728Z
#> 54 54   2017-0034       <NA>      10    NA 2018-08-15T19:52:41.841Z
#> 55 55   2017-0049 2017-09-15      14    NA 2018-08-15T19:52:41.958Z
#> 56 56   2016-0011 2016-07-12      15    NA 2018-08-15T19:52:42.067Z
#> 57 57   2016-0017 2016-07-25      16    NA 2018-08-15T19:52:42.178Z
#> 58 58   2016-0023 2016-07-20      17    NA 2018-08-15T19:52:42.294Z
#> 59 59   2016-0029 2016-07-28      18    NA 2018-08-15T19:52:42.407Z
#> 60 60   2016-0005 2016-07-21      19    NA 2018-08-15T19:52:42.517Z
#> 61 61   2017-0041 2017-08-21      25    NA 2018-08-15T19:52:42.630Z
#> 62 62   2017-0029 2017-07-11      26    NA 2018-08-15T19:52:42.741Z
#> 63 63   2017-0023 2017-07-10      21    NA 2018-08-15T19:52:42.876Z
#> 64 64   2017-0017 2017-07-11      27    NA 2018-08-15T19:52:42.988Z
#> 65 65   2017-0005 2017-06-29      22    NA 2018-08-15T19:52:43.107Z
#> 66 66   2017-0011 2017-07-10      23    NA 2018-08-15T19:52:43.214Z
#> 67 67   2017-0035       <NA>      24    NA 2018-08-15T19:52:43.323Z
#> 68 68   2017-0050 2017-09-15      28    NA 2018-08-15T19:52:43.441Z
#> 69 69   2016-0012 2016-07-12      29    NA 2018-08-15T19:52:43.547Z
#> 70 70   2016-0018 2016-07-25      30    NA 2018-08-15T19:52:43.661Z
#> 71 71   2016-0024 2016-07-20      31    NA 2018-08-15T19:52:43.778Z
#> 72 72   2016-0030 2016-07-28      32    NA 2018-08-15T19:52:43.890Z
#> 73 73   2016-0006 2016-07-21      33    NA 2018-08-15T19:52:43.999Z
#> 74 74   2017-0042 2017-08-21      39    NA 2018-08-15T19:52:44.115Z
#> 75 75   2017-0030 2017-07-11      40    NA 2018-08-15T19:52:44.257Z
#> 76 76   2017-0024 2017-07-10      35    NA 2018-08-15T19:52:44.368Z
#> 77 77   2017-0018 2017-07-11      41    NA 2018-08-15T19:52:44.482Z
#> 78 78   2017-0006 2017-06-29      36    NA 2018-08-15T19:52:44.592Z
#> 79 79   2017-0012 2017-07-10      37    NA 2018-08-15T19:52:44.702Z
#> 80 80   2017-0036       <NA>      38    NA 2018-08-15T19:52:44.812Z
#> 81 81   2017-0051 2017-09-15      42    NA 2018-08-15T19:52:44.941Z
#>                  updated_at
#> 1  2018-08-15T19:52:35.626Z
#> 2  2018-08-15T19:52:35.739Z
#> 3  2018-08-15T19:52:35.863Z
#> 4  2018-08-15T19:52:35.973Z
#> 5  2018-08-15T19:52:36.086Z
#> 6  2018-08-15T19:52:36.196Z
#> 7  2018-08-15T19:52:36.313Z
#> 8  2018-08-15T19:52:36.457Z
#> 9  2018-08-15T19:52:36.570Z
#> 10 2018-08-15T19:52:36.689Z
#> 11 2018-08-15T19:52:36.801Z
#> 12 2018-08-15T19:52:36.910Z
#> 13 2018-08-15T19:52:37.053Z
#> 14 2018-08-15T19:52:37.160Z
#> 15 2018-08-15T19:52:37.271Z
#> 16 2018-08-15T19:52:37.385Z
#> 17 2018-08-15T19:52:37.499Z
#> 18 2018-08-15T19:52:37.612Z
#> 19 2018-08-15T19:52:37.758Z
#> 20 2018-08-15T19:52:37.874Z
#> 21 2018-08-15T19:52:37.986Z
#> 22 2018-08-15T19:52:38.099Z
#> 23 2018-08-15T19:52:38.225Z
#> 24 2018-08-15T19:52:38.336Z
#> 25 2018-08-15T19:52:38.446Z
#> 26 2018-08-15T19:52:38.566Z
#> 27 2018-08-15T19:52:38.682Z
#> 28 2018-08-15T19:52:38.791Z
#> 29 2018-08-15T19:52:38.911Z
#> 30 2018-08-15T19:52:39.050Z
#> 31 2018-08-15T19:52:39.165Z
#> 32 2018-08-15T19:52:39.282Z
#> 33 2018-08-15T19:52:39.394Z
#> 34 2018-08-15T19:52:39.502Z
#> 35 2018-08-15T19:52:39.615Z
#> 36 2018-08-15T19:52:39.734Z
#> 37 2018-08-15T19:52:39.844Z
#> 38 2018-08-15T19:52:39.960Z
#> 39 2018-08-15T19:52:40.079Z
#> 40 2018-08-15T19:52:40.190Z
#> 41 2018-08-15T19:52:40.330Z
#> 42 2018-08-15T19:52:40.450Z
#> 43 2018-08-15T19:52:40.560Z
#> 44 2018-08-15T19:52:40.674Z
#> 45 2018-08-15T19:52:40.789Z
#> 46 2018-08-15T19:52:40.906Z
#> 47 2018-08-15T19:52:41.016Z
#> 48 2018-08-15T19:52:41.128Z
#> 49 2018-08-15T19:52:41.248Z
#> 50 2018-08-15T19:52:41.359Z
#> 51 2018-08-15T19:52:41.474Z
#> 52 2018-08-15T19:52:41.619Z
#> 53 2018-08-15T19:52:41.728Z
#> 54 2018-08-15T19:52:41.841Z
#> 55 2018-08-15T19:52:41.958Z
#> 56 2018-08-15T19:52:42.067Z
#> 57 2018-08-15T19:52:42.178Z
#> 58 2018-08-15T19:52:42.294Z
#> 59 2018-08-15T19:52:42.407Z
#> 60 2018-08-15T19:52:42.517Z
#> 61 2018-08-15T19:52:42.630Z
#> 62 2018-08-15T19:52:42.741Z
#> 63 2018-08-15T19:52:42.876Z
#> 64 2018-08-15T19:52:42.988Z
#> 65 2018-08-15T19:52:43.107Z
#> 66 2018-08-15T19:52:43.214Z
#> 67 2018-08-15T19:52:43.323Z
#> 68 2018-08-15T19:52:43.441Z
#> 69 2018-08-15T19:52:43.547Z
#> 70 2018-08-15T19:52:43.661Z
#> 71 2018-08-15T19:52:43.778Z
#> 72 2018-08-15T19:52:43.890Z
#> 73 2018-08-15T19:52:43.999Z
#> 74 2018-08-15T19:52:44.115Z
#> 75 2018-08-15T19:52:44.257Z
#> 76 2018-08-15T19:52:44.368Z
#> 77 2018-08-15T19:52:44.482Z
#> 78 2018-08-15T19:52:44.592Z
#> 79 2018-08-15T19:52:44.702Z
#> 80 2018-08-15T19:52:44.812Z
#> 81 2018-08-15T19:52:44.941Z
#> 
#> $response
#> Response [http://localhost:3001/api/v1/samples?page=0]
#>   Date: 2018-08-15 21:57
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 13.5 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"

get_samples()

#> [[1]]
#> [[1]][[1]]
#> $body
#>    id sample_code  date_samp trap_id notes               created_at
#> 1   1   2016-0007 2016-06-10       1    NA 2018-08-15T19:52:35.626Z
#> 2   2   2016-0013 2016-06-28       2    NA 2018-08-15T19:52:35.739Z
#> 3   3   2016-0019 2016-06-20       3    NA 2018-08-15T19:52:35.863Z
#> 4   4   2016-0025 2016-07-11       4    NA 2018-08-15T19:52:35.973Z
#> 5   5   2016-0001 2016-06-21       5    NA 2018-08-15T19:52:36.086Z
#> 6   6   2017-0037 2017-07-28      11    NA 2018-08-15T19:52:36.196Z
#> 7   7   2017-0043 2017-07-11       6    NA 2018-08-15T19:52:36.313Z
#> 8   8   2017-0025 2017-06-15      12    NA 2018-08-15T19:52:36.457Z
#> 9   9   2017-0019 2017-06-20       7    NA 2018-08-15T19:52:36.570Z
#> 10 10   2017-0013 2017-06-14      13    NA 2018-08-15T19:52:36.689Z
#> 11 11   2017-0001 2017-06-08       8    NA 2018-08-15T19:52:36.801Z
#> 12 12   2017-0007 2017-06-20       9    NA 2018-08-15T19:52:36.910Z
#> 13 13   2017-0031       <NA>      10    NA 2018-08-15T19:52:37.053Z
#> 14 14   2017-0046 2017-09-02      14    NA 2018-08-15T19:52:37.160Z
#> 15 15   2016-0008 2016-06-10      15    NA 2018-08-15T19:52:37.271Z
#> 16 16   2016-0014 2016-06-28      16    NA 2018-08-15T19:52:37.385Z
#> 17 17   2016-0020 2016-06-20      17    NA 2018-08-15T19:52:37.499Z
#> 18 18   2016-0026 2016-07-11      18    NA 2018-08-15T19:52:37.612Z
#> 19 19   2016-0002 2016-06-21      19    NA 2018-08-15T19:52:37.758Z
#> 20 20   2017-0038 2017-07-28      25    NA 2018-08-15T19:52:37.874Z
#> 21 21   2017-0044 2017-07-11      20    NA 2018-08-15T19:52:37.986Z
#> 22 22   2017-0026 2017-06-15      26    NA 2018-08-15T19:52:38.099Z
#> 23 23   2017-0020 2017-06-20      21    NA 2018-08-15T19:52:38.225Z
#> 24 24   2017-0014 2017-06-14      27    NA 2018-08-15T19:52:38.336Z
#> 25 25   2017-0002 2017-06-08      22    NA 2018-08-15T19:52:38.446Z
#> 26 26   2017-0008 2017-06-20      23    NA 2018-08-15T19:52:38.566Z
#> 27 27   2017-0032       <NA>      24    NA 2018-08-15T19:52:38.682Z
#> 28 28   2017-0047 2017-09-02      28    NA 2018-08-15T19:52:38.791Z
#> 29 29   2016-0009 2016-06-10      29    NA 2018-08-15T19:52:38.911Z
#> 30 30   2016-0015 2016-06-28      30    NA 2018-08-15T19:52:39.050Z
#> 31 31   2016-0021 2016-06-20      31    NA 2018-08-15T19:52:39.165Z
#> 32 32   2016-0027 2016-07-11      32    NA 2018-08-15T19:52:39.282Z
#> 33 33   2016-0003 2016-06-21      33    NA 2018-08-15T19:52:39.394Z
#> 34 34   2017-0039 2017-07-28      39    NA 2018-08-15T19:52:39.502Z
#> 35 35   2017-0045 2017-07-11      34    NA 2018-08-15T19:52:39.615Z
#> 36 36   2017-0027 2017-06-15      40    NA 2018-08-15T19:52:39.734Z
#> 37 37   2017-0021 2017-06-20      35    NA 2018-08-15T19:52:39.844Z
#> 38 38   2017-0015 2017-06-14      41    NA 2018-08-15T19:52:39.960Z
#> 39 39   2017-0003 2017-06-08      36    NA 2018-08-15T19:52:40.079Z
#> 40 40   2017-0009 2017-06-20      37    NA 2018-08-15T19:52:40.190Z
#> 41 41   2017-0033       <NA>      38    NA 2018-08-15T19:52:40.330Z
#> 42 42   2017-0048 2017-09-02      42    NA 2018-08-15T19:52:40.450Z
#> 43 43   2016-0010 2016-07-12       1    NA 2018-08-15T19:52:40.560Z
#> 44 44   2016-0016 2016-07-25       2    NA 2018-08-15T19:52:40.674Z
#> 45 45   2016-0022 2016-07-20       3    NA 2018-08-15T19:52:40.789Z
#> 46 46   2016-0028 2016-07-28       4    NA 2018-08-15T19:52:40.906Z
#> 47 47   2016-0004 2016-07-21       5    NA 2018-08-15T19:52:41.016Z
#> 48 48   2017-0040 2017-08-21      11    NA 2018-08-15T19:52:41.128Z
#> 49 49   2017-0028 2017-07-11      12    NA 2018-08-15T19:52:41.248Z
#> 50 50   2017-0022 2017-07-10       7    NA 2018-08-15T19:52:41.359Z
#> 51 51   2017-0016 2017-07-11      13    NA 2018-08-15T19:52:41.474Z
#> 52 52   2017-0004 2017-06-29       8    NA 2018-08-15T19:52:41.619Z
#> 53 53   2017-0010 2017-07-10       9    NA 2018-08-15T19:52:41.728Z
#> 54 54   2017-0034       <NA>      10    NA 2018-08-15T19:52:41.841Z
#> 55 55   2017-0049 2017-09-15      14    NA 2018-08-15T19:52:41.958Z
#> 56 56   2016-0011 2016-07-12      15    NA 2018-08-15T19:52:42.067Z
#> 57 57   2016-0017 2016-07-25      16    NA 2018-08-15T19:52:42.178Z
#> 58 58   2016-0023 2016-07-20      17    NA 2018-08-15T19:52:42.294Z
#> 59 59   2016-0029 2016-07-28      18    NA 2018-08-15T19:52:42.407Z
#> 60 60   2016-0005 2016-07-21      19    NA 2018-08-15T19:52:42.517Z
#> 61 61   2017-0041 2017-08-21      25    NA 2018-08-15T19:52:42.630Z
#> 62 62   2017-0029 2017-07-11      26    NA 2018-08-15T19:52:42.741Z
#> 63 63   2017-0023 2017-07-10      21    NA 2018-08-15T19:52:42.876Z
#> 64 64   2017-0017 2017-07-11      27    NA 2018-08-15T19:52:42.988Z
#> 65 65   2017-0005 2017-06-29      22    NA 2018-08-15T19:52:43.107Z
#> 66 66   2017-0011 2017-07-10      23    NA 2018-08-15T19:52:43.214Z
#> 67 67   2017-0035       <NA>      24    NA 2018-08-15T19:52:43.323Z
#> 68 68   2017-0050 2017-09-15      28    NA 2018-08-15T19:52:43.441Z
#> 69 69   2016-0012 2016-07-12      29    NA 2018-08-15T19:52:43.547Z
#> 70 70   2016-0018 2016-07-25      30    NA 2018-08-15T19:52:43.661Z
#> 71 71   2016-0024 2016-07-20      31    NA 2018-08-15T19:52:43.778Z
#> 72 72   2016-0030 2016-07-28      32    NA 2018-08-15T19:52:43.890Z
#> 73 73   2016-0006 2016-07-21      33    NA 2018-08-15T19:52:43.999Z
#> 74 74   2017-0042 2017-08-21      39    NA 2018-08-15T19:52:44.115Z
#> 75 75   2017-0030 2017-07-11      40    NA 2018-08-15T19:52:44.257Z
#> 76 76   2017-0024 2017-07-10      35    NA 2018-08-15T19:52:44.368Z
#> 77 77   2017-0018 2017-07-11      41    NA 2018-08-15T19:52:44.482Z
#> 78 78   2017-0006 2017-06-29      36    NA 2018-08-15T19:52:44.592Z
#> 79 79   2017-0012 2017-07-10      37    NA 2018-08-15T19:52:44.702Z
#> 80 80   2017-0036       <NA>      38    NA 2018-08-15T19:52:44.812Z
#> 81 81   2017-0051 2017-09-15      42    NA 2018-08-15T19:52:44.941Z
#>                  updated_at
#> 1  2018-08-15T19:52:35.626Z
#> 2  2018-08-15T19:52:35.739Z
#> 3  2018-08-15T19:52:35.863Z
#> 4  2018-08-15T19:52:35.973Z
#> 5  2018-08-15T19:52:36.086Z
#> 6  2018-08-15T19:52:36.196Z
#> 7  2018-08-15T19:52:36.313Z
#> 8  2018-08-15T19:52:36.457Z
#> 9  2018-08-15T19:52:36.570Z
#> 10 2018-08-15T19:52:36.689Z
#> 11 2018-08-15T19:52:36.801Z
#> 12 2018-08-15T19:52:36.910Z
#> 13 2018-08-15T19:52:37.053Z
#> 14 2018-08-15T19:52:37.160Z
#> 15 2018-08-15T19:52:37.271Z
#> 16 2018-08-15T19:52:37.385Z
#> 17 2018-08-15T19:52:37.499Z
#> 18 2018-08-15T19:52:37.612Z
#> 19 2018-08-15T19:52:37.758Z
#> 20 2018-08-15T19:52:37.874Z
#> 21 2018-08-15T19:52:37.986Z
#> 22 2018-08-15T19:52:38.099Z
#> 23 2018-08-15T19:52:38.225Z
#> 24 2018-08-15T19:52:38.336Z
#> 25 2018-08-15T19:52:38.446Z
#> 26 2018-08-15T19:52:38.566Z
#> 27 2018-08-15T19:52:38.682Z
#> 28 2018-08-15T19:52:38.791Z
#> 29 2018-08-15T19:52:38.911Z
#> 30 2018-08-15T19:52:39.050Z
#> 31 2018-08-15T19:52:39.165Z
#> 32 2018-08-15T19:52:39.282Z
#> 33 2018-08-15T19:52:39.394Z
#> 34 2018-08-15T19:52:39.502Z
#> 35 2018-08-15T19:52:39.615Z
#> 36 2018-08-15T19:52:39.734Z
#> 37 2018-08-15T19:52:39.844Z
#> 38 2018-08-15T19:52:39.960Z
#> 39 2018-08-15T19:52:40.079Z
#> 40 2018-08-15T19:52:40.190Z
#> 41 2018-08-15T19:52:40.330Z
#> 42 2018-08-15T19:52:40.450Z
#> 43 2018-08-15T19:52:40.560Z
#> 44 2018-08-15T19:52:40.674Z
#> 45 2018-08-15T19:52:40.789Z
#> 46 2018-08-15T19:52:40.906Z
#> 47 2018-08-15T19:52:41.016Z
#> 48 2018-08-15T19:52:41.128Z
#> 49 2018-08-15T19:52:41.248Z
#> 50 2018-08-15T19:52:41.359Z
#> 51 2018-08-15T19:52:41.474Z
#> 52 2018-08-15T19:52:41.619Z
#> 53 2018-08-15T19:52:41.728Z
#> 54 2018-08-15T19:52:41.841Z
#> 55 2018-08-15T19:52:41.958Z
#> 56 2018-08-15T19:52:42.067Z
#> 57 2018-08-15T19:52:42.178Z
#> 58 2018-08-15T19:52:42.294Z
#> 59 2018-08-15T19:52:42.407Z
#> 60 2018-08-15T19:52:42.517Z
#> 61 2018-08-15T19:52:42.630Z
#> 62 2018-08-15T19:52:42.741Z
#> 63 2018-08-15T19:52:42.876Z
#> 64 2018-08-15T19:52:42.988Z
#> 65 2018-08-15T19:52:43.107Z
#> 66 2018-08-15T19:52:43.214Z
#> 67 2018-08-15T19:52:43.323Z
#> 68 2018-08-15T19:52:43.441Z
#> 69 2018-08-15T19:52:43.547Z
#> 70 2018-08-15T19:52:43.661Z
#> 71 2018-08-15T19:52:43.778Z
#> 72 2018-08-15T19:52:43.890Z
#> 73 2018-08-15T19:52:43.999Z
#> 74 2018-08-15T19:52:44.115Z
#> 75 2018-08-15T19:52:44.257Z
#> 76 2018-08-15T19:52:44.368Z
#> 77 2018-08-15T19:52:44.482Z
#> 78 2018-08-15T19:52:44.592Z
#> 79 2018-08-15T19:52:44.702Z
#> 80 2018-08-15T19:52:44.812Z
#> 81 2018-08-15T19:52:44.941Z
#> 
#> $response
#> Response [http://localhost:3001/api/v1/samples?page=0]
#>   Date: 2018-08-15 21:57
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 13.5 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"