Obtenir les informations sur les sites depuis l'API de coleo

Obtenir les informations sur les sites depuis l'API de coleo

get_sites(site_code = NULL, ...)

Arguments

site_code
character vecteur contenant les identifiants uniques que l'on désire obtenir. Si site_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_sites(site_code=c('135_104_H01'))

#> [[1]]
#> [[1]][[1]]
#> $body
#>   id cell_id off_station_code_id   site_code   type  opened_at notes
#> 1  1     134                  NA 135_104_H01 marais 2016-06-23    NA
#>                 created_at               updated_at
#> 1 2018-08-16T17:25:11.237Z 2018-08-16T17:25:11.237Z
#>                                                                                                                                                                                                                                                                                                                                                                                                               campaigns
#> 1 1, 27, 83, 1, 1, 1, végétation, papilionidés, odonates, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, NA, NA, NA, 2018-08-16T17:25:12.793Z, 2018-08-16T17:26:09.930Z, 2018-08-16T17:28:21.730Z, 2018-08-16T17:25:12.793Z, 2018-08-16T17:26:09.930Z, 2018-08-16T17:28:21.730Z
#>   geom.type  geom.coordinates
#> 1     Point 46.8088, -72.2994
#> 
#> $response
#> Response [http://localhost:3001/api/v1/sites?site_code=135_104_H01&page=0]
#>   Date: 2018-08-16 18:40
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 1.01 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"

get_sites()

#> [[1]]
#> [[1]][[1]]
#> $body
#>    id cell_id off_station_code_id   site_code          type  opened_at notes
#> 1  42     495                <NA>  80_175_T01    toundrique 2017-08-18    NA
#> 2  41     446                <NA>  48_181_T01    toundrique 2017-07-13    NA
#> 3  26     141                <NA> 136_116_T01    toundrique 2017-07-20    NA
#> 4  25     149          46577142-1 137_107_F01     forestier 2017-05-29    NA
#> 5  24     160          47117049-2 139_111_F01     forestier 2017-05-25    NA
#> 6  23     157          46157157-2 139_103_F01     forestier 2017-05-23    NA
#> 7  22     169          46427042-2 142_111_F01     forestier 2017-05-23    NA
#> 8  21     186          45527111-2 145_102_F01     forestier 2017-05-18    NA
#> 9  20     186                <NA> 145_102_H02        marais 2017-07-13    NA
#> 10 19     186                <NA> 145_102_H01     tourbière 2017-07-13    NA
#> 11 18     157                <NA> 139_103_H02 marais côtier 2017-07-04    NA
#> 12 17     169                <NA> 142_111_H01     tourbière 2017-07-05    NA
#> 13 16     169                <NA> 142_111_H02        marais 2017-07-05    NA
#> 14 15     157                <NA> 139_103_H01     tourbière 2017-06-28    NA
#> 15 14     160                <NA> 139_111_H01     tourbière 2017-06-29    NA
#> 16 13     149                <NA> 137_107_H03 marais côtier 2017-06-26    NA
#> 17 12     149                <NA> 137_107_H02        marais 2017-06-22    NA
#> 18 11     149                <NA> 137_107_H01     tourbière 2017-06-22    NA
#> 19 10     119                <NA> 132_116_H01     tourbière 2016-06-28    NA
#> 20  9     119          48137116-1 132_116_F01     forestier 2016-06-03    NA
#> 21  8     165                <NA> 141_108_H01     tourbière 2016-07-05    NA
#> 22  7     165          46407102-1 141_108_F01     forestier 2016-05-26    NA
#> 23  6     151      471827710948-1 137_111_F01     forestier 2016-06-14    NA
#> 24  5     198                <NA> 148_101_H01     tourbière 2016-06-21    NA
#> 25  4     151                <NA> 137_111_H01     tourbière 2016-07-11    NA
#> 26  3     198                <NA> 148_101_F01     forestier 2016-05-17    NA
#> 27  2     134          46487218-2 135_104_F01     forestier 2016-05-20    NA
#> 28  1     134                <NA> 135_104_H01        marais 2016-06-23    NA
#>                  created_at               updated_at
#> 1  2018-08-16T17:26:26.152Z 2018-08-16T17:26:26.152Z
#> 2  2018-08-16T17:26:26.089Z 2018-08-16T17:26:26.089Z
#> 3  2018-08-16T17:25:12.637Z 2018-08-16T17:25:12.637Z
#> 4  2018-08-16T17:25:12.586Z 2018-08-16T17:25:12.586Z
#> 5  2018-08-16T17:25:12.534Z 2018-08-16T17:25:12.534Z
#> 6  2018-08-16T17:25:12.486Z 2018-08-16T17:25:12.486Z
#> 7  2018-08-16T17:25:12.439Z 2018-08-16T17:25:12.439Z
#> 8  2018-08-16T17:25:12.390Z 2018-08-16T17:25:12.390Z
#> 9  2018-08-16T17:25:12.337Z 2018-08-16T17:25:12.337Z
#> 10 2018-08-16T17:25:12.283Z 2018-08-16T17:25:12.283Z
#> 11 2018-08-16T17:25:12.231Z 2018-08-16T17:25:12.231Z
#> 12 2018-08-16T17:25:12.177Z 2018-08-16T17:25:12.177Z
#> 13 2018-08-16T17:25:12.128Z 2018-08-16T17:25:12.128Z
#> 14 2018-08-16T17:25:12.078Z 2018-08-16T17:25:12.078Z
#> 15 2018-08-16T17:25:12.028Z 2018-08-16T17:25:12.028Z
#> 16 2018-08-16T17:25:11.977Z 2018-08-16T17:25:11.977Z
#> 17 2018-08-16T17:25:11.928Z 2018-08-16T17:25:11.928Z
#> 18 2018-08-16T17:25:11.868Z 2018-08-16T17:25:11.868Z
#> 19 2018-08-16T17:25:11.790Z 2018-08-16T17:25:11.790Z
#> 20 2018-08-16T17:25:11.737Z 2018-08-16T17:25:11.737Z
#> 21 2018-08-16T17:25:11.685Z 2018-08-16T17:25:11.685Z
#> 22 2018-08-16T17:25:11.623Z 2018-08-16T17:25:11.623Z
#> 23 2018-08-16T17:25:11.573Z 2018-08-16T17:25:11.573Z
#> 24 2018-08-16T17:25:11.476Z 2018-08-16T17:25:11.476Z
#> 25 2018-08-16T17:25:11.420Z 2018-08-16T17:25:11.420Z
#> 26 2018-08-16T17:25:11.362Z 2018-08-16T17:25:11.362Z
#> 27 2018-08-16T17:25:11.296Z 2018-08-16T17:25:11.296Z
#> 28 2018-08-16T17:25:11.237Z 2018-08-16T17:25:11.237Z
#>                                                                                                                                                                                                                                                                                                                                                                                                                    campaigns
#> 1                                                                                                                                                                                                                                                                                                        54, 42, microfaunes, Francis Demers, 2017-08-18, 2017-09-15, NA, 2018-08-16T17:26:27.067Z, 2018-08-16T17:26:27.067Z
#> 2                                                                                                                                                                                                                                                                                                         51, 41, microfaunes, Samuel Gagnon, 2017-07-13, 2017-08-21, NA, 2018-08-16T17:26:26.895Z, 2018-08-16T17:26:26.895Z
#> 3                         26, 50, 34, 26, 26, 26, végétation, microfaunes, papilionidés, Caroline Dubé, Adine Séguin, Guillaume Tremblay, Caroline Dubé, Caroline Dubé, Adine Séguin, 2017-07-20, 2017-06-07, 2017-07-20, 2017-07-20, NA, 2017-07-20, NA, NA, NA, 2018-08-16T17:25:14.663Z, 2018-08-16T17:26:26.831Z, 2018-08-16T17:26:10.546Z, 2018-08-16T17:25:14.663Z, 2018-08-16T17:26:26.831Z, 2018-08-16T17:26:10.546Z
#> 4                                                                                                                                              25, 49, 25, 25, végétation, microfaunes, Caroline Dubé, Guillaume Tremblay, Guillaume Tremblay, Caroline Dubé, 2017-05-29, 2017-05-29, 2017-07-10, 2017-07-10, NA, NA, 2018-08-16T17:25:14.579Z, 2018-08-16T17:26:26.777Z, 2018-08-16T17:25:14.579Z, 2018-08-16T17:26:26.777Z
#> 5                                                                                                                                              24, 47, 24, 24, végétation, microfaunes, Caroline Dubé, Guillaume Tremblay, Guillaume Tremblay, Caroline Dubé, 2017-05-25, 2017-05-25, 2017-07-10, 2017-07-10, NA, NA, 2018-08-16T17:25:14.517Z, 2018-08-16T17:26:26.660Z, 2018-08-16T17:25:14.517Z, 2018-08-16T17:26:26.660Z
#> 6                                                                                                                                                                  23, 52, 23, 23, végétation, microfaunes, Caroline Dubé, Guillaume Tremblay, Caroline Dubé, 2017-05-23, 2017-05-23, 2017-07-11, 2017-07-11, NA, NA, 2018-08-16T17:25:14.454Z, 2018-08-16T17:26:26.960Z, 2018-08-16T17:25:14.454Z, 2018-08-16T17:26:26.960Z
#> 7      53, 46, 22, 22, 22, 22, microfaunes, microfaunes, végétation, Caroline Dubé, Anouk Simard, Anouk Simard, Caroline Dubé, Guillaume Tremblay, Caroline Dubé, Guillaume Tremblay, 2017-05-23, 2017-06-14, 2017-05-23, 2017-07-11, NA, 2017-07-10, NA, NA, NA, 2018-08-16T17:26:27.012Z, 2018-08-16T17:26:26.607Z, 2018-08-16T17:25:14.392Z, 2018-08-16T17:26:27.012Z, 2018-08-16T17:26:26.607Z, 2018-08-16T17:25:14.392Z
#> 8                                                                                                                                                    21, 48, 21, 21, végétation, microfaunes, Caroline Dubé, Stéphane Poulin, Stéphane Poulin, Caroline Dubé, 2017-05-18, 2017-05-18, 2017-07-13, 2017-06-29, NA, NA, 2018-08-16T17:25:14.320Z, 2018-08-16T17:26:26.715Z, 2018-08-16T17:25:14.320Z, 2018-08-16T17:26:26.715Z
#> 9                                                                                                                                                                  32, 20, 20, 20, papilionidés, végétation, Stéphane Poulin, Caroline Dubé, Stéphane Poulin, 2017-07-12, 2017-07-13, 2017-07-12, 2017-07-13, NA, NA, 2018-08-16T17:26:10.374Z, 2018-08-16T17:25:14.259Z, 2018-08-16T17:26:10.374Z, 2018-08-16T17:25:14.259Z
#> 10                                                                                                                                                                 19, 33, 19, 19, végétation, papilionidés, Caroline Dubé, Stéphane Poulin, Stéphane Poulin, 2017-07-13, 2017-07-06, 2017-07-13, 2017-07-06, NA, NA, 2018-08-16T17:25:14.198Z, 2018-08-16T17:26:10.448Z, 2018-08-16T17:25:14.198Z, 2018-08-16T17:26:10.448Z
#> 11                                                                                                                                                                                                                                                                                     18, 18, végétation, Caroline Dubé, Guillaume Tremblay, 2017-07-04, 2017-07-04, NA, 2018-08-16T17:25:14.132Z, 2018-08-16T17:25:14.132Z
#> 12                                                                                                                                                36, 17, 17, 17, papilionidés, végétation, Caroline Dubé, Émilie Trépanier, Caroline Dubé, Émilie Trépanier, 2017-07-05, 2017-07-05, 2017-07-05, 2017-07-05, NA, NA, 2018-08-16T17:26:10.711Z, 2018-08-16T17:25:14.071Z, 2018-08-16T17:26:10.711Z, 2018-08-16T17:25:14.071Z
#> 13                                                                                                                                                16, 35, 16, 16, végétation, papilionidés, Caroline Dubé, Émilie Trépanier, Caroline Dubé, Émilie Trépanier, 2017-07-05, 2017-07-05, 2017-07-05, 2017-07-05, NA, NA, 2018-08-16T17:25:14.010Z, 2018-08-16T17:26:10.636Z, 2018-08-16T17:25:14.010Z, 2018-08-16T17:26:10.636Z
#> 14                           38, 15, 88, 15, 15, 15, papilionidés, végétation, odonates, Caroline Dubé, Émilie Trépanier, Caroline Dubé, Émilie Trépanier, Caroline Dubé, NA, 2017-06-28, 2017-06-28, 2017-06-28, 2017-06-28, 2017-06-28, 2017-06-28, NA, NA, NA, 2018-08-16T17:26:10.880Z, 2018-08-16T17:25:13.945Z, 2018-08-16T17:28:22.219Z, 2018-08-16T17:26:10.880Z, 2018-08-16T17:25:13.945Z, 2018-08-16T17:28:22.219Z
#> 15                                                                                                                                            14, 37, 14, 14, végétation, papilionidés, Caroline Dubé, Guillaume Tremblay, Caroline Dubé, Guillaume Tremblay, 2017-06-29, 2017-06-29, 2017-06-29, 2017-06-29, NA, NA, 2018-08-16T17:25:13.884Z, 2018-08-16T17:26:10.793Z, 2018-08-16T17:25:13.884Z, 2018-08-16T17:26:10.793Z
#> 16                                                                                                                                                                                                                                                                                           13, 13, végétation, Caroline Dubé, Anouk Simard, 2017-06-26, 2017-06-26, NA, 2018-08-16T17:25:13.820Z, 2018-08-16T17:25:13.820Z
#> 17                                                                                                                                            40, 12, 12, 12, papilionidés, végétation, Caroline Dubé, Guillaume Tremblay, Caroline Dubé, Guillaume Tremblay, 2017-06-22, 2017-06-22, 2017-06-22, 2017-06-22, NA, NA, 2018-08-16T17:26:11.053Z, 2018-08-16T17:25:13.755Z, 2018-08-16T17:26:11.053Z, 2018-08-16T17:25:13.755Z
#> 18                                                                                                                                                  39, 11, 11, 11, papilionidés, végétation, Caroline Dubé, Anouk Simard, Caroline Dubé, Guillaume Tremblay, 2017-06-26, 2017-06-22, 2017-06-26, 2017-06-22, NA, NA, 2018-08-16T17:26:10.953Z, 2018-08-16T17:25:13.695Z, 2018-08-16T17:26:10.953Z, 2018-08-16T17:25:13.695Z
#> 19 10, 28, 84, 10, 10, 10, végétation, papilionidés, odonates, Caroline Dubé, Pierre-Alexis Drolet, Pierre-Alexis Drolet, Caroline Dubé, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-28, 2016-06-28, 2016-06-28, 2016-06-28, 2016-06-28, 2016-06-28, NA, NA, NA, 2018-08-16T17:25:13.633Z, 2018-08-16T17:26:10.017Z, 2018-08-16T17:28:21.804Z, 2018-08-16T17:25:13.633Z, 2018-08-16T17:26:10.017Z, 2018-08-16T17:28:21.804Z
#> 20                                                                                                                                            42, 9, 9, 9, microfaunes, végétation, Pierre-Alexis Drolet, Caroline Dubé, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-03, 2016-06-03, 2016-07-25, 2016-07-25, NA, NA, 2018-08-16T17:26:26.363Z, 2018-08-16T17:25:13.564Z, 2018-08-16T17:26:26.363Z, 2018-08-16T17:25:13.564Z
#> 21                    29, 8, 85, 8, 8, 8, papilionidés, végétation, odonates, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-07-05, 2016-07-05, 2016-07-05, 2016-07-05, 2016-07-05, 2016-07-05, NA, NA, NA, 2018-08-16T17:26:10.099Z, 2018-08-16T17:25:13.487Z, 2018-08-16T17:28:21.900Z, 2018-08-16T17:26:10.099Z, 2018-08-16T17:25:13.487Z, 2018-08-16T17:28:21.900Z
#> 22                                                                                                                                            7, 43, 7, 7, végétation, microfaunes, Caroline Dubé, Pierre-Alexis Drolet, Pierre-Alexis Drolet, Caroline Dubé, 2016-05-26, 2016-05-26, 2016-07-20, 2016-07-20, NA, NA, 2018-08-16T17:25:13.386Z, 2018-08-16T17:26:26.418Z, 2018-08-16T17:25:13.386Z, 2018-08-16T17:26:26.418Z
#> 23                                                                                                                                            44, 6, 6, 6, microfaunes, végétation, Pierre-Alexis Drolet, Caroline Dubé, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-14, 2016-06-14, 2016-07-28, 2016-07-28, NA, NA, 2018-08-16T17:26:26.493Z, 2018-08-16T17:25:13.175Z, 2018-08-16T17:26:26.493Z, 2018-08-16T17:25:13.175Z
#> 24     5, 31, 87, 5, 5, 5, végétation, papilionidés, odonates, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-21, 2016-06-30, 2016-06-30, 2016-06-21, 2016-06-30, 2016-06-30, NA, NA, NA, 2018-08-16T17:25:13.102Z, 2018-08-16T17:26:10.289Z, 2018-08-16T17:28:22.121Z, 2018-08-16T17:25:13.102Z, 2018-08-16T17:26:10.289Z, 2018-08-16T17:28:22.121Z
#> 25     30, 86, 4, 4, 4, 4, papilionidés, odonates, végétation, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-07-11, 2016-07-11, 2016-07-11, 2016-07-11, 2016-07-11, 2016-07-11, NA, NA, NA, 2018-08-16T17:26:10.171Z, 2018-08-16T17:28:22.012Z, 2018-08-16T17:25:13.036Z, 2018-08-16T17:26:10.171Z, 2018-08-16T17:28:22.012Z, 2018-08-16T17:25:13.036Z
#> 26                                                                                                                          3, 45, 3, 3, végétation, microfaunes, Caroline Dubé, Antoine St-Louis, Pierre-Alexis Drolet, Pierre-Alexis Drolet, Caroline Dubé, 2016-05-17, 2016-05-17, 2016-07-21, 2016-07-21, NA, NA, 2018-08-16T17:25:12.957Z, 2018-08-16T17:26:26.546Z, 2018-08-16T17:25:12.957Z, 2018-08-16T17:26:26.546Z
#> 27                                                                                                                                            2, 41, 2, 2, végétation, microfaunes, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-05-20, 2016-05-20, 2016-07-12, 2016-07-12, NA, NA, 2018-08-16T17:25:12.890Z, 2018-08-16T17:26:26.288Z, 2018-08-16T17:25:12.890Z, 2018-08-16T17:26:26.288Z
#> 28     1, 27, 83, 1, 1, 1, végétation, papilionidés, odonates, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, NA, NA, NA, 2018-08-16T17:25:12.793Z, 2018-08-16T17:26:09.930Z, 2018-08-16T17:28:21.730Z, 2018-08-16T17:25:12.793Z, 2018-08-16T17:26:09.930Z, 2018-08-16T17:28:21.730Z
#>    geom.type  geom.coordinates
#> 1      Point 58.6462, -69.9970
#> 2      Point 62.2018, -75.6395
#> 3      Point 47.6666, -70.7795
#> 4      Point 46.9580, -71.6948
#> 5      Point 47.1785, -70.8145
#> 6      Point 46.2502, -71.9497
#> 7      Point 46.7112, -70.7035
#> 8      Point 45.8590, -71.1879
#> 9      Point 45.8420, -71.1738
#> 10     Point 45.9644, -71.1378
#> 11     Point 46.6590, -71.8344
#> 12     Point 46.8893, -70.4352
#> 13     Point 46.8707, -70.4730
#> 14     Point 46.3908, -71.8162
#> 15     Point 47.1199, -70.7453
#> 16     Point 46.7312, -71.4364
#> 17     Point 46.8684, -71.6753
#> 18     Point 46.8766, -71.6651
#> 19     Point 48.2158, -71.2814
#> 20     Point 48.2163, -71.2666
#> 21     Point 46.7822, -71.0368
#> 22     Point 46.6722, -71.0299
#> 23     Point 47.3074, -71.1624
#> 24     Point 45.4475, -70.8866
#> 25     Point 47.2572, -71.1631
#> 26     Point 45.4599, -71.0001
#> 27     Point 46.7941, -72.3030
#> 28     Point 46.8088, -72.2994
#> 
#> $response
#> Response [http://localhost:3001/api/v1/sites?page=0]
#>   Date: 2018-08-16 18:40
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 22.4 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"