Obtenir les informations sur les sites depuis l'API de coleo

Obtenir les informations sur les sites depuis l'API de coleo

get_sites(site_code = NULL, ...)

Arguments

site_code
character vecteur contenant les identifiants uniques que l'on désire obtenir. Si site_code n'est pas spécifié, la fonction retournera l'ensemble des entrées présentes dans la table cells.

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_sites(site_code=c('135_104_H01'))

#> [[1]]
#> [[1]]$body
#>   id cell_id off_station_code_id   site_code   type  opened_at notes
#> 1  1     134                  NA 135_104_H01 marais 2016-06-23    NA
#>                 created_at               updated_at
#> 1 2018-08-10T17:15:22.839Z 2018-08-10T17:15:22.839Z
#>                                                                                                                                                                                                                                                                                                                                                                                                               campaigns
#> 1 1, 27, 83, 1, 1, 1, végétation, papilionidés, odonates, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, NA, NA, NA, 2018-08-10T17:15:32.558Z, 2018-08-10T17:17:51.647Z, 2018-08-10T17:28:08.629Z, 2018-08-10T17:15:32.558Z, 2018-08-10T17:17:51.647Z, 2018-08-10T17:28:08.629Z
#>   geom.type  geom.coordinates
#> 1     Point 46.8088, -72.2994
#> 
#> [[1]]$response
#> Response [http://localhost:3001/api/v1/sites?site_code=135_104_H01&page=0]
#>   Date: 2018-08-13 15:28
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 1.01 kB
#> 
#> 
#> 

get_sites()

#> [[1]]
#> [[1]][[1]]
#> $body
#>    id cell_id off_station_code_id   site_code          type  opened_at notes
#> 1  42     495                <NA>  80_175_T01    toundrique 2017-08-18    NA
#> 2  41     446                <NA>  48_181_T01    toundrique 2017-07-13    NA
#> 3  26     141                <NA> 136_116_T01    toundrique 2017-07-20    NA
#> 4  25     149          46577142-1 137_107_F01     forestier 2017-05-29    NA
#> 5  24     160          47117049-2 139_111_F01     forestier 2017-05-25    NA
#> 6  23     157          46157157-2 139_103_F01     forestier 2017-05-23    NA
#> 7  22     169          46427042-2 142_111_F01     forestier 2017-05-23    NA
#> 8  21     186          45527111-2 145_102_F01     forestier 2017-05-18    NA
#> 9  20     186                <NA> 145_102_H02        marais 2017-07-13    NA
#> 10 19     186                <NA> 145_102_H01     tourbière 2017-07-13    NA
#> 11 18     157                <NA> 139_103_H02 marais côtier 2017-07-04    NA
#> 12 17     169                <NA> 142_111_H01     tourbière 2017-07-05    NA
#> 13 16     169                <NA> 142_111_H02        marais 2017-07-05    NA
#> 14 15     157                <NA> 139_103_H01     tourbière 2017-06-28    NA
#> 15 14     160                <NA> 139_111_H01     tourbière 2017-06-29    NA
#> 16 13     149                <NA> 137_107_H03 marais côtier 2017-06-26    NA
#> 17 12     149                <NA> 137_107_H02        marais 2017-06-22    NA
#> 18 11     149                <NA> 137_107_H01     tourbière 2017-06-22    NA
#> 19 10     119                <NA> 132_116_H01     tourbière 2016-06-28    NA
#> 20  9     119          48137116-1 132_116_F01     forestier 2016-06-03    NA
#> 21  8     165                <NA> 141_108_H01     tourbière 2016-07-05    NA
#> 22  7     165          46407102-1 141_108_F01     forestier 2016-05-26    NA
#> 23  6     151      471827710948-1 137_111_F01     forestier 2016-06-14    NA
#> 24  5     198                <NA> 148_101_H01     tourbière 2016-06-21    NA
#> 25  4     151                <NA> 137_111_H01     tourbière 2016-07-11    NA
#> 26  3     198                <NA> 148_101_F01     forestier 2016-05-17    NA
#> 27  2     134          46487218-2 135_104_F01     forestier 2016-05-20    NA
#> 28  1     134                <NA> 135_104_H01        marais 2016-06-23    NA
#>                  created_at               updated_at
#> 1  2018-08-10T17:19:57.249Z 2018-08-10T17:19:57.249Z
#> 2  2018-08-10T17:19:57.175Z 2018-08-10T17:19:57.175Z
#> 3  2018-08-10T17:15:24.556Z 2018-08-10T17:15:24.556Z
#> 4  2018-08-10T17:15:24.489Z 2018-08-10T17:15:24.489Z
#> 5  2018-08-10T17:15:24.400Z 2018-08-10T17:15:24.400Z
#> 6  2018-08-10T17:15:24.281Z 2018-08-10T17:15:24.281Z
#> 7  2018-08-10T17:15:24.159Z 2018-08-10T17:15:24.159Z
#> 8  2018-08-10T17:15:24.069Z 2018-08-10T17:15:24.069Z
#> 9  2018-08-10T17:15:24.001Z 2018-08-10T17:15:24.001Z
#> 10 2018-08-10T17:15:23.952Z 2018-08-10T17:15:23.952Z
#> 11 2018-08-10T17:15:23.907Z 2018-08-10T17:15:23.907Z
#> 12 2018-08-10T17:15:23.861Z 2018-08-10T17:15:23.861Z
#> 13 2018-08-10T17:15:23.780Z 2018-08-10T17:15:23.780Z
#> 14 2018-08-10T17:15:23.733Z 2018-08-10T17:15:23.733Z
#> 15 2018-08-10T17:15:23.675Z 2018-08-10T17:15:23.675Z
#> 16 2018-08-10T17:15:23.616Z 2018-08-10T17:15:23.616Z
#> 17 2018-08-10T17:15:23.559Z 2018-08-10T17:15:23.559Z
#> 18 2018-08-10T17:15:23.500Z 2018-08-10T17:15:23.500Z
#> 19 2018-08-10T17:15:23.436Z 2018-08-10T17:15:23.436Z
#> 20 2018-08-10T17:15:23.363Z 2018-08-10T17:15:23.363Z
#> 21 2018-08-10T17:15:23.297Z 2018-08-10T17:15:23.297Z
#> 22 2018-08-10T17:15:23.215Z 2018-08-10T17:15:23.215Z
#> 23 2018-08-10T17:15:23.155Z 2018-08-10T17:15:23.155Z
#> 24 2018-08-10T17:15:23.095Z 2018-08-10T17:15:23.095Z
#> 25 2018-08-10T17:15:23.022Z 2018-08-10T17:15:23.022Z
#> 26 2018-08-10T17:15:22.969Z 2018-08-10T17:15:22.969Z
#> 27 2018-08-10T17:15:22.916Z 2018-08-10T17:15:22.916Z
#> 28 2018-08-10T17:15:22.839Z 2018-08-10T17:15:22.839Z
#>                                                                                                                                                                                                                                                                                                                                                                                                                    campaigns
#> 1                                                                                                                                                                                                                                                                                                        54, 42, microfaunes, Francis Demers, 2017-08-18, 2017-09-15, NA, 2018-08-10T17:19:58.269Z, 2018-08-10T17:19:58.269Z
#> 2                                                                                                                                                                                                                                                                                                         51, 41, microfaunes, Samuel Gagnon, 2017-07-13, 2017-08-21, NA, 2018-08-10T17:19:58.044Z, 2018-08-10T17:19:58.044Z
#> 3                         26, 50, 34, 26, 26, 26, végétation, microfaunes, papilionidés, Caroline Dubé, Adine Séguin, Guillaume Tremblay, Caroline Dubé, Caroline Dubé, Adine Séguin, 2017-07-20, 2017-06-07, 2017-07-20, 2017-07-20, NA, 2017-07-20, NA, NA, NA, 2018-08-10T17:15:50.074Z, 2018-08-10T17:19:57.973Z, 2018-08-10T17:17:52.101Z, 2018-08-10T17:15:50.074Z, 2018-08-10T17:19:57.973Z, 2018-08-10T17:17:52.101Z
#> 4                                                                                                                                              25, 49, 25, 25, végétation, microfaunes, Caroline Dubé, Guillaume Tremblay, Guillaume Tremblay, Caroline Dubé, 2017-05-29, 2017-05-29, 2017-07-10, 2017-07-10, NA, NA, 2018-08-10T17:15:49.986Z, 2018-08-10T17:19:57.921Z, 2018-08-10T17:15:49.986Z, 2018-08-10T17:19:57.921Z
#> 5                                                                                                                                              24, 47, 24, 24, végétation, microfaunes, Caroline Dubé, Guillaume Tremblay, Guillaume Tremblay, Caroline Dubé, 2017-05-25, 2017-05-25, 2017-07-10, 2017-07-10, NA, NA, 2018-08-10T17:15:49.899Z, 2018-08-10T17:19:57.806Z, 2018-08-10T17:15:49.899Z, 2018-08-10T17:19:57.806Z
#> 6                                                                                                                                                                  23, 52, 23, 23, végétation, microfaunes, Caroline Dubé, Guillaume Tremblay, Caroline Dubé, 2017-05-23, 2017-05-23, 2017-07-11, 2017-07-11, NA, NA, 2018-08-10T17:15:49.807Z, 2018-08-10T17:19:58.105Z, 2018-08-10T17:15:49.807Z, 2018-08-10T17:19:58.105Z
#> 7      53, 46, 22, 22, 22, 22, microfaunes, microfaunes, végétation, Caroline Dubé, Anouk Simard, Anouk Simard, Caroline Dubé, Guillaume Tremblay, Caroline Dubé, Guillaume Tremblay, 2017-05-23, 2017-06-14, 2017-05-23, 2017-07-11, NA, 2017-07-10, NA, NA, NA, 2018-08-10T17:19:58.212Z, 2018-08-10T17:19:57.738Z, 2018-08-10T17:15:49.706Z, 2018-08-10T17:19:58.212Z, 2018-08-10T17:19:57.738Z, 2018-08-10T17:15:49.706Z
#> 8                                                                                                                                                    21, 48, 21, 21, végétation, microfaunes, Caroline Dubé, Stéphane Poulin, Stéphane Poulin, Caroline Dubé, 2017-05-18, 2017-05-18, 2017-07-13, 2017-06-29, NA, NA, 2018-08-10T17:15:49.638Z, 2018-08-10T17:19:57.865Z, 2018-08-10T17:15:49.638Z, 2018-08-10T17:19:57.865Z
#> 9                                                                                                                                                                  32, 20, 20, 20, papilionidés, végétation, Stéphane Poulin, Caroline Dubé, Stéphane Poulin, 2017-07-12, 2017-07-13, 2017-07-12, 2017-07-13, NA, NA, 2018-08-10T17:17:51.978Z, 2018-08-10T17:15:49.569Z, 2018-08-10T17:17:51.978Z, 2018-08-10T17:15:49.569Z
#> 10                                                                                                                                                                 19, 33, 19, 19, végétation, papilionidés, Caroline Dubé, Stéphane Poulin, Stéphane Poulin, 2017-07-13, 2017-07-06, 2017-07-13, 2017-07-06, NA, NA, 2018-08-10T17:15:49.494Z, 2018-08-10T17:17:52.041Z, 2018-08-10T17:15:49.494Z, 2018-08-10T17:17:52.041Z
#> 11                                                                                                                                                                                                                                                                                     18, 18, végétation, Caroline Dubé, Guillaume Tremblay, 2017-07-04, 2017-07-04, NA, 2018-08-10T17:15:49.408Z, 2018-08-10T17:15:49.408Z
#> 12                                                                                                                                                36, 17, 17, 17, papilionidés, végétation, Caroline Dubé, Émilie Trépanier, Caroline Dubé, Émilie Trépanier, 2017-07-05, 2017-07-05, 2017-07-05, 2017-07-05, NA, NA, 2018-08-10T17:17:52.230Z, 2018-08-10T17:15:49.306Z, 2018-08-10T17:17:52.230Z, 2018-08-10T17:15:49.306Z
#> 13                                                                                                                                                16, 35, 16, 16, végétation, papilionidés, Caroline Dubé, Émilie Trépanier, Caroline Dubé, Émilie Trépanier, 2017-07-05, 2017-07-05, 2017-07-05, 2017-07-05, NA, NA, 2018-08-10T17:15:49.208Z, 2018-08-10T17:17:52.168Z, 2018-08-10T17:15:49.208Z, 2018-08-10T17:17:52.168Z
#> 14                           38, 15, 88, 15, 15, 15, papilionidés, végétation, odonates, Caroline Dubé, Émilie Trépanier, Caroline Dubé, Émilie Trépanier, Caroline Dubé, NA, 2017-06-28, 2017-06-28, 2017-06-28, 2017-06-28, 2017-06-28, 2017-06-28, NA, NA, NA, 2018-08-10T17:17:52.390Z, 2018-08-10T17:15:49.103Z, 2018-08-10T17:28:09.051Z, 2018-08-10T17:17:52.390Z, 2018-08-10T17:15:49.103Z, 2018-08-10T17:28:09.051Z
#> 15                                                                                                                                            14, 37, 14, 14, végétation, papilionidés, Caroline Dubé, Guillaume Tremblay, Caroline Dubé, Guillaume Tremblay, 2017-06-29, 2017-06-29, 2017-06-29, 2017-06-29, NA, NA, 2018-08-10T17:15:48.964Z, 2018-08-10T17:17:52.331Z, 2018-08-10T17:15:48.964Z, 2018-08-10T17:17:52.331Z
#> 16                                                                                                                                                                                                                                                                                           13, 13, végétation, Caroline Dubé, Anouk Simard, 2017-06-26, 2017-06-26, NA, 2018-08-10T17:15:48.843Z, 2018-08-10T17:15:48.843Z
#> 17                                                                                                                                            40, 12, 12, 12, papilionidés, végétation, Caroline Dubé, Guillaume Tremblay, Caroline Dubé, Guillaume Tremblay, 2017-06-22, 2017-06-22, 2017-06-22, 2017-06-22, NA, NA, 2018-08-10T17:17:52.509Z, 2018-08-10T17:15:48.730Z, 2018-08-10T17:17:52.509Z, 2018-08-10T17:15:48.730Z
#> 18                                                                                                                                                  39, 11, 11, 11, papilionidés, végétation, Caroline Dubé, Anouk Simard, Caroline Dubé, Guillaume Tremblay, 2017-06-26, 2017-06-22, 2017-06-26, 2017-06-22, NA, NA, 2018-08-10T17:17:52.447Z, 2018-08-10T17:15:48.559Z, 2018-08-10T17:17:52.447Z, 2018-08-10T17:15:48.559Z
#> 19 10, 28, 84, 10, 10, 10, végétation, papilionidés, odonates, Caroline Dubé, Pierre-Alexis Drolet, Pierre-Alexis Drolet, Caroline Dubé, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-28, 2016-06-28, 2016-06-28, 2016-06-28, 2016-06-28, 2016-06-28, NA, NA, NA, 2018-08-10T17:15:48.423Z, 2018-08-10T17:17:51.717Z, 2018-08-10T17:28:08.714Z, 2018-08-10T17:15:48.423Z, 2018-08-10T17:17:51.717Z, 2018-08-10T17:28:08.714Z
#> 20                                                                                                                                            42, 9, 9, 9, microfaunes, végétation, Pierre-Alexis Drolet, Caroline Dubé, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-03, 2016-06-03, 2016-07-25, 2016-07-25, NA, NA, 2018-08-10T17:19:57.462Z, 2018-08-10T17:15:48.303Z, 2018-08-10T17:19:57.462Z, 2018-08-10T17:15:48.303Z
#> 21                    29, 8, 85, 8, 8, 8, papilionidés, végétation, odonates, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-07-05, 2016-07-05, 2016-07-05, 2016-07-05, 2016-07-05, 2016-07-05, NA, NA, NA, 2018-08-10T17:17:51.783Z, 2018-08-10T17:15:48.186Z, 2018-08-10T17:28:08.776Z, 2018-08-10T17:17:51.783Z, 2018-08-10T17:15:48.186Z, 2018-08-10T17:28:08.776Z
#> 22                                                                                                                                            7, 43, 7, 7, végétation, microfaunes, Caroline Dubé, Pierre-Alexis Drolet, Pierre-Alexis Drolet, Caroline Dubé, 2016-05-26, 2016-05-26, 2016-07-20, 2016-07-20, NA, NA, 2018-08-10T17:15:48.079Z, 2018-08-10T17:19:57.533Z, 2018-08-10T17:15:48.079Z, 2018-08-10T17:19:57.533Z
#> 23                                                                                                                                            44, 6, 6, 6, microfaunes, végétation, Pierre-Alexis Drolet, Caroline Dubé, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-14, 2016-06-14, 2016-07-28, 2016-07-28, NA, NA, 2018-08-10T17:19:57.606Z, 2018-08-10T17:15:47.982Z, 2018-08-10T17:19:57.606Z, 2018-08-10T17:15:47.982Z
#> 24     5, 31, 87, 5, 5, 5, végétation, papilionidés, odonates, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-21, 2016-06-30, 2016-06-30, 2016-06-21, 2016-06-30, 2016-06-30, NA, NA, NA, 2018-08-10T17:15:47.905Z, 2018-08-10T17:17:51.921Z, 2018-08-10T17:28:08.972Z, 2018-08-10T17:15:47.905Z, 2018-08-10T17:17:51.921Z, 2018-08-10T17:28:08.972Z
#> 25     30, 86, 4, 4, 4, 4, papilionidés, odonates, végétation, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-07-11, 2016-07-11, 2016-07-11, 2016-07-11, 2016-07-11, 2016-07-11, NA, NA, NA, 2018-08-10T17:17:51.860Z, 2018-08-10T17:28:08.895Z, 2018-08-10T17:15:47.778Z, 2018-08-10T17:17:51.860Z, 2018-08-10T17:28:08.895Z, 2018-08-10T17:15:47.778Z
#> 26                                                                                                                          3, 45, 3, 3, végétation, microfaunes, Caroline Dubé, Antoine St-Louis, Pierre-Alexis Drolet, Pierre-Alexis Drolet, Caroline Dubé, 2016-05-17, 2016-05-17, 2016-07-21, 2016-07-21, NA, NA, 2018-08-10T17:15:47.673Z, 2018-08-10T17:19:57.662Z, 2018-08-10T17:15:47.673Z, 2018-08-10T17:19:57.662Z
#> 27                                                                                                                                            2, 41, 2, 2, végétation, microfaunes, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-05-20, 2016-05-20, 2016-07-12, 2016-07-12, NA, NA, 2018-08-10T17:15:47.547Z, 2018-08-10T17:19:57.398Z, 2018-08-10T17:15:47.547Z, 2018-08-10T17:19:57.398Z
#> 28     1, 27, 83, 1, 1, 1, végétation, papilionidés, odonates, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, Caroline Dubé, Pierre-Alexis Drolet, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, 2016-06-23, NA, NA, NA, 2018-08-10T17:15:32.558Z, 2018-08-10T17:17:51.647Z, 2018-08-10T17:28:08.629Z, 2018-08-10T17:15:32.558Z, 2018-08-10T17:17:51.647Z, 2018-08-10T17:28:08.629Z
#>    geom.type  geom.coordinates
#> 1      Point 58.6462, -69.9970
#> 2      Point 62.2018, -75.6395
#> 3      Point 47.6666, -70.7795
#> 4      Point 46.9580, -71.6948
#> 5      Point 47.1785, -70.8145
#> 6      Point 46.2502, -71.9497
#> 7      Point 46.7112, -70.7035
#> 8      Point 45.8590, -71.1879
#> 9      Point 45.8420, -71.1738
#> 10     Point 45.9644, -71.1378
#> 11     Point 46.6590, -71.8344
#> 12     Point 46.8893, -70.4352
#> 13     Point 46.8707, -70.4730
#> 14     Point 46.3908, -71.8162
#> 15     Point 47.1199, -70.7453
#> 16     Point 46.7312, -71.4364
#> 17     Point 46.8684, -71.6753
#> 18     Point 46.8766, -71.6651
#> 19     Point 48.2158, -71.2814
#> 20     Point 48.2163, -71.2666
#> 21     Point 46.7822, -71.0368
#> 22     Point 46.6722, -71.0299
#> 23     Point 47.3074, -71.1624
#> 24     Point 45.4475, -70.8866
#> 25     Point 47.2572, -71.1631
#> 26     Point 45.4599, -71.0001
#> 27     Point 46.7941, -72.3030
#> 28     Point 46.8088, -72.2994
#> 
#> $response
#> Response [http://localhost:3001/api/v1/sites?page=0]
#>   Date: 2018-08-13 15:28
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 22.4 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#>