Retourne les entrées pour les pièges présents dans la base de données

Retourne les entrées pour les pièges présents dans la base de données

get_traps(trap_code = NULL, campaign_id = NULL, ...)

Arguments

trap_code
character vecteur contenant les codes de trappes désirés

campaign_id
integer vecteur contenant les identifiants uniques de la campagne

...
httr options; arguments de la fonction httr::GET()

Value

Retourne une objet de type list contenant les réponses de l'API. Chaque niveau de la liste correspond à une page. Pour chacun des appels sur l'API (page), la classe retourné est getSuccess ou getError. Une réponse de classe getSuccess est une liste à deux niveaux composé du contenu (body), et la réponse httr::response. Une réponse de classe getError dispose de la même structure mais ne contiendra pas de body, seulement la réponse de l'API.

Examples

get_traps(trap_code=c('MP-22'))

#> [[1]]
#> [[1]][[1]]
#> $body
#>   id trap_code campaign_id notes               created_at
#> 1 11     MP-22          51    NA 2018-08-16T17:26:32.335Z
#>                 updated_at
#> 1 2018-08-16T17:26:32.335Z
#>                                                                                                                            landmarks
#> 1 30, 51, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.340Z, 2018-08-16T17:26:32.340Z, 11, NA, NA, Point, 62.2018, -75.6396
#>                                                                                                                                                                       samples
#> 1 48, 6, 2017-0040, 2017-0037, 2017-08-21, 2017-07-28, 11, 11, NA, NA, 2018-08-16T17:26:42.906Z, 2018-08-16T17:26:36.750Z, 2018-08-16T17:26:42.906Z, 2018-08-16T17:26:36.750Z
#> 
#> $response
#> Response [http://localhost:3001/api/v1/traps?trap_code=MP-22&page=0]
#>   Date: 2018-08-16 18:40
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 838 B
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"

get_traps()

#> [[1]]
#> [[1]][[1]]
#> $body
#>    id trap_code campaign_id notes               created_at
#> 1   1     MP-05          41    NA 2018-08-16T17:26:31.373Z
#> 2   2     MP-13          42    NA 2018-08-16T17:26:31.493Z
#> 3   3     MP-10          43    NA 2018-08-16T17:26:31.586Z
#> 4   4     MP-03          44    NA 2018-08-16T17:26:31.675Z
#> 5   5     MP-04          45    NA 2018-08-16T17:26:31.761Z
#> 6   6    MPL-01          46    NA 2018-08-16T17:26:31.850Z
#> 7   7     MP-13          47    NA 2018-08-16T17:26:31.977Z
#> 8   8     MP-01          48    NA 2018-08-16T17:26:32.071Z
#> 9   9     MP-06          49    NA 2018-08-16T17:26:32.157Z
#> 10 10     MP-21          50    NA 2018-08-16T17:26:32.242Z
#> 11 11     MP-22          51    NA 2018-08-16T17:26:32.335Z
#> 12 12     MP-11          52    NA 2018-08-16T17:26:32.432Z
#> 13 13     MP-10          53    NA 2018-08-16T17:26:32.527Z
#> 14 14     MP-16          54    NA 2018-08-16T17:26:32.612Z
#> 15 15     MP-06          41    NA 2018-08-16T17:26:32.701Z
#> 16 16     MP-12          42    NA 2018-08-16T17:26:32.795Z
#> 17 17     MP-09          43    NA 2018-08-16T17:26:32.884Z
#> 18 18     MP-11          44    NA 2018-08-16T17:26:32.974Z
#> 19 19     MP-01          45    NA 2018-08-16T17:26:33.070Z
#> 20 20    MPL-03          46    NA 2018-08-16T17:26:33.157Z
#> 21 21     MP-09          47    NA 2018-08-16T17:26:33.250Z
#> 22 22     MP-02          48    NA 2018-08-16T17:26:33.370Z
#> 23 23     MP-05          49    NA 2018-08-16T17:26:33.456Z
#> 24 24     MP-20          50    NA 2018-08-16T17:26:33.553Z
#> 25 25     MP-29          51    NA 2018-08-16T17:26:33.643Z
#> 26 26     MP-04          52    NA 2018-08-16T17:26:33.738Z
#> 27 27     MP-08          53    NA 2018-08-16T17:26:33.823Z
#> 28 28     MP-18          54    NA 2018-08-16T17:26:33.921Z
#> 29 29     MP-07          41    NA 2018-08-16T17:26:34.053Z
#> 30 30     MP-14          42    NA 2018-08-16T17:26:34.164Z
#> 31 31     MP-08          43    NA 2018-08-16T17:26:34.266Z
#> 32 32     MP-15          44    NA 2018-08-16T17:26:34.361Z
#> 33 33     MP-02          45    NA 2018-08-16T17:26:34.460Z
#> 34 34    MPL-02          46    NA 2018-08-16T17:26:34.610Z
#> 35 35     MP-12          47    NA 2018-08-16T17:26:34.756Z
#> 36 36     MP-03          48    NA 2018-08-16T17:26:34.891Z
#> 37 37     MP-15          49    NA 2018-08-16T17:26:35.035Z
#> 38 38     MP-17          50    NA 2018-08-16T17:26:35.143Z
#> 39 39     MP-23          51    NA 2018-08-16T17:26:35.272Z
#> 40 40     MP-14          52    NA 2018-08-16T17:26:35.414Z
#> 41 41     MP-07          53    NA 2018-08-16T17:26:35.551Z
#> 42 42     MP-19          54    NA 2018-08-16T17:26:35.681Z
#>                  updated_at
#> 1  2018-08-16T17:26:31.373Z
#> 2  2018-08-16T17:26:31.493Z
#> 3  2018-08-16T17:26:31.586Z
#> 4  2018-08-16T17:26:31.675Z
#> 5  2018-08-16T17:26:31.761Z
#> 6  2018-08-16T17:26:31.850Z
#> 7  2018-08-16T17:26:31.977Z
#> 8  2018-08-16T17:26:32.071Z
#> 9  2018-08-16T17:26:32.157Z
#> 10 2018-08-16T17:26:32.242Z
#> 11 2018-08-16T17:26:32.335Z
#> 12 2018-08-16T17:26:32.432Z
#> 13 2018-08-16T17:26:32.527Z
#> 14 2018-08-16T17:26:32.612Z
#> 15 2018-08-16T17:26:32.701Z
#> 16 2018-08-16T17:26:32.795Z
#> 17 2018-08-16T17:26:32.884Z
#> 18 2018-08-16T17:26:32.974Z
#> 19 2018-08-16T17:26:33.070Z
#> 20 2018-08-16T17:26:33.157Z
#> 21 2018-08-16T17:26:33.250Z
#> 22 2018-08-16T17:26:33.370Z
#> 23 2018-08-16T17:26:33.456Z
#> 24 2018-08-16T17:26:33.553Z
#> 25 2018-08-16T17:26:33.643Z
#> 26 2018-08-16T17:26:33.738Z
#> 27 2018-08-16T17:26:33.823Z
#> 28 2018-08-16T17:26:33.921Z
#> 29 2018-08-16T17:26:34.053Z
#> 30 2018-08-16T17:26:34.164Z
#> 31 2018-08-16T17:26:34.266Z
#> 32 2018-08-16T17:26:34.361Z
#> 33 2018-08-16T17:26:34.460Z
#> 34 2018-08-16T17:26:34.610Z
#> 35 2018-08-16T17:26:34.756Z
#> 36 2018-08-16T17:26:34.891Z
#> 37 2018-08-16T17:26:35.035Z
#> 38 2018-08-16T17:26:35.143Z
#> 39 2018-08-16T17:26:35.272Z
#> 40 2018-08-16T17:26:35.414Z
#> 41 2018-08-16T17:26:35.551Z
#> 42 2018-08-16T17:26:35.681Z
#>                                                                                                                             landmarks
#> 1    20, 41, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:31.378Z, 2018-08-16T17:26:31.378Z, 1, NA, NA, Point, 46.7942, -72.303
#> 2   21, 42, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:31.498Z, 2018-08-16T17:26:31.498Z, 2, NA, NA, Point, 48.2164, -71.2668
#> 3   22, 43, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:31.591Z, 2018-08-16T17:26:31.591Z, 3, NA, NA, Point, 46.6723, -71.0297
#> 4   23, 44, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:31.679Z, 2018-08-16T17:26:31.679Z, 4, NA, NA, Point, 47.3073, -71.1623
#> 5   24, 45, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:31.766Z, 2018-08-16T17:26:31.766Z, 5, NA, NA, Point, 45.4599, -71.0002
#> 6    25, 46, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:31.855Z, 2018-08-16T17:26:31.855Z, 6, NA, NA, Point, 46.711, -70.7036
#> 7   26, 47, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:31.982Z, 2018-08-16T17:26:31.982Z, 7, NA, NA, Point, 47.1785, -70.8147
#> 8    27, 48, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.074Z, 2018-08-16T17:26:32.074Z, 8, NA, NA, Point, 45.859, -71.1879
#> 9   28, 49, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.161Z, 2018-08-16T17:26:32.161Z, 9, NA, NA, Point, 46.9579, -71.6948
#> 10 29, 50, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.247Z, 2018-08-16T17:26:32.247Z, 10, NA, NA, Point, 47.6665, -70.7793
#> 11 30, 51, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.340Z, 2018-08-16T17:26:32.340Z, 11, NA, NA, Point, 62.2018, -75.6396
#> 12 31, 52, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.438Z, 2018-08-16T17:26:32.438Z, 12, NA, NA, Point, 46.2503, -71.9498
#> 13 32, 53, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.531Z, 2018-08-16T17:26:32.531Z, 13, NA, NA, Point, 46.7111, -70.7035
#> 14 33, 54, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.617Z, 2018-08-16T17:26:32.617Z, 14, NA, NA, Point, 58.6461, -69.9971
#> 15  34, 41, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.705Z, 2018-08-16T17:26:32.705Z, 15, NA, NA, Point, 46.794, -72.3031
#> 16 35, 42, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.800Z, 2018-08-16T17:26:32.800Z, 16, NA, NA, Point, 48.2164, -71.2666
#> 17 36, 43, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.888Z, 2018-08-16T17:26:32.888Z, 17, NA, NA, Point, 46.7822, -71.0368
#> 18 37, 44, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:32.979Z, 2018-08-16T17:26:32.979Z, 18, NA, NA, Point, 47.3073, -71.1622
#> 19 38, 45, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.074Z, 2018-08-16T17:26:33.074Z, 19, NA, NA, Point, 45.4599, -71.0001
#> 20  39, 46, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.161Z, 2018-08-16T17:26:33.161Z, 20, NA, NA, Point, 46.711, -70.7034
#> 21 40, 47, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.255Z, 2018-08-16T17:26:33.255Z, 21, NA, NA, Point, 47.1786, -70.8144
#> 22  41, 48, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.373Z, 2018-08-16T17:26:33.373Z, 22, NA, NA, Point, 45.859, -71.1878
#> 23 42, 49, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.461Z, 2018-08-16T17:26:33.461Z, 23, NA, NA, Point, 46.9581, -71.6949
#> 24 43, 50, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.557Z, 2018-08-16T17:26:33.557Z, 24, NA, NA, Point, 47.6667, -70.7793
#> 25 44, 51, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.647Z, 2018-08-16T17:26:33.647Z, 25, NA, NA, Point, 62.2018, -75.6395
#> 26 45, 52, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.743Z, 2018-08-16T17:26:33.743Z, 26, NA, NA, Point, 46.2502, -71.9498
#> 27 46, 53, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.828Z, 2018-08-16T17:26:33.828Z, 27, NA, NA, Point, 46.7113, -70.7035
#> 28  47, 54, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:33.926Z, 2018-08-16T17:26:33.926Z, 28, NA, NA, Point, 58.6462, -69.997
#> 29 48, 41, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:34.058Z, 2018-08-16T17:26:34.058Z, 29, NA, NA, Point, 46.7941, -72.3034
#> 30 49, 42, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:34.168Z, 2018-08-16T17:26:34.168Z, 30, NA, NA, Point, 48.2163, -71.2664
#> 31 50, 43, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:34.271Z, 2018-08-16T17:26:34.271Z, 31, NA, NA, Point, 46.6722, -71.0299
#> 32 51, 44, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:34.365Z, 2018-08-16T17:26:34.365Z, 32, NA, NA, Point, 47.3075, -71.1624
#> 33 52, 45, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:34.468Z, 2018-08-16T17:26:34.468Z, 33, NA, NA, Point, 45.4599, -70.9999
#> 34 53, 46, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:34.617Z, 2018-08-16T17:26:34.617Z, 34, NA, NA, Point, 46.7114, -70.7038
#> 35 54, 47, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:34.769Z, 2018-08-16T17:26:34.769Z, 35, NA, NA, Point, 47.1786, -70.8147
#> 36   55, 48, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:34.896Z, 2018-08-16T17:26:34.896Z, 36, NA, NA, Point, 45.859, -71.188
#> 37  56, 49, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:35.040Z, 2018-08-16T17:26:35.040Z, 37, NA, NA, Point, 46.958, -71.6945
#> 38 57, 50, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:35.148Z, 2018-08-16T17:26:35.148Z, 38, NA, NA, Point, 47.6665, -70.7791
#> 39 58, 51, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:35.278Z, 2018-08-16T17:26:35.278Z, 39, NA, NA, Point, 62.2018, -75.6393
#> 40 59, 52, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:35.420Z, 2018-08-16T17:26:35.420Z, 40, NA, NA, Point, 46.2504, -71.9497
#> 41 60, 53, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:35.557Z, 2018-08-16T17:26:35.557Z, 41, NA, NA, Point, 46.7114, -70.7034
#> 42  61, 54, NA, NA, 0, mm, 0, 0, cm, gps, NA, 2018-08-16T17:26:35.687Z, 2018-08-16T17:26:35.687Z, 42, NA, NA, Point, 58.6463, -69.997
#>                                                                                                                                                                         samples
#> 1     43, 1, 2016-0010, 2016-0007, 2016-07-12, 2016-06-10, 1, 1, NA, NA, 2018-08-16T17:26:42.184Z, 2018-08-16T17:26:36.029Z, 2018-08-16T17:26:42.184Z, 2018-08-16T17:26:36.029Z
#> 2     44, 2, 2016-0016, 2016-0013, 2016-07-25, 2016-06-28, 2, 2, NA, NA, 2018-08-16T17:26:42.305Z, 2018-08-16T17:26:36.172Z, 2018-08-16T17:26:42.305Z, 2018-08-16T17:26:36.172Z
#> 3     45, 3, 2016-0022, 2016-0019, 2016-07-20, 2016-06-20, 3, 3, NA, NA, 2018-08-16T17:26:42.428Z, 2018-08-16T17:26:36.330Z, 2018-08-16T17:26:42.428Z, 2018-08-16T17:26:36.330Z
#> 4     46, 4, 2016-0028, 2016-0025, 2016-07-28, 2016-07-11, 4, 4, NA, NA, 2018-08-16T17:26:42.590Z, 2018-08-16T17:26:36.496Z, 2018-08-16T17:26:42.590Z, 2018-08-16T17:26:36.496Z
#> 5     47, 5, 2016-0004, 2016-0001, 2016-07-21, 2016-06-21, 5, 5, NA, NA, 2018-08-16T17:26:42.739Z, 2018-08-16T17:26:36.619Z, 2018-08-16T17:26:42.739Z, 2018-08-16T17:26:36.619Z
#> 6                                                                                           7, 2017-0043, 2017-07-11, 6, NA, 2018-08-16T17:26:36.898Z, 2018-08-16T17:26:36.898Z
#> 7     50, 9, 2017-0022, 2017-0019, 2017-07-10, 2017-06-20, 7, 7, NA, NA, 2018-08-16T17:26:43.349Z, 2018-08-16T17:26:37.147Z, 2018-08-16T17:26:43.349Z, 2018-08-16T17:26:37.147Z
#> 8    52, 11, 2017-0004, 2017-0001, 2017-06-29, 2017-06-08, 8, 8, NA, NA, 2018-08-16T17:26:43.764Z, 2018-08-16T17:26:37.416Z, 2018-08-16T17:26:43.764Z, 2018-08-16T17:26:37.416Z
#> 9    53, 12, 2017-0010, 2017-0007, 2017-07-10, 2017-06-20, 9, 9, NA, NA, 2018-08-16T17:26:43.964Z, 2018-08-16T17:26:37.542Z, 2018-08-16T17:26:43.964Z, 2018-08-16T17:26:37.542Z
#> 10                 54, 13, 2017-0034, 2017-0031, NA, NA, 10, 10, NA, NA, 2018-08-16T17:26:44.112Z, 2018-08-16T17:26:37.665Z, 2018-08-16T17:26:44.112Z, 2018-08-16T17:26:37.665Z
#> 11  48, 6, 2017-0040, 2017-0037, 2017-08-21, 2017-07-28, 11, 11, NA, NA, 2018-08-16T17:26:42.906Z, 2018-08-16T17:26:36.750Z, 2018-08-16T17:26:42.906Z, 2018-08-16T17:26:36.750Z
#> 12  49, 8, 2017-0028, 2017-0025, 2017-07-11, 2017-06-15, 12, 12, NA, NA, 2018-08-16T17:26:43.135Z, 2018-08-16T17:26:37.022Z, 2018-08-16T17:26:43.135Z, 2018-08-16T17:26:37.022Z
#> 13 51, 10, 2017-0016, 2017-0013, 2017-07-11, 2017-06-14, 13, 13, NA, NA, 2018-08-16T17:26:43.552Z, 2018-08-16T17:26:37.274Z, 2018-08-16T17:26:43.552Z, 2018-08-16T17:26:37.274Z
#> 14 55, 14, 2017-0049, 2017-0046, 2017-09-15, 2017-09-02, 14, 14, NA, NA, 2018-08-16T17:26:44.256Z, 2018-08-16T17:26:37.785Z, 2018-08-16T17:26:44.256Z, 2018-08-16T17:26:37.785Z
#> 15 56, 15, 2016-0011, 2016-0008, 2016-07-12, 2016-06-10, 15, 15, NA, NA, 2018-08-16T17:26:44.398Z, 2018-08-16T17:26:37.938Z, 2018-08-16T17:26:44.398Z, 2018-08-16T17:26:37.938Z
#> 16 57, 16, 2016-0017, 2016-0014, 2016-07-25, 2016-06-28, 16, 16, NA, NA, 2018-08-16T17:26:44.529Z, 2018-08-16T17:26:38.102Z, 2018-08-16T17:26:44.529Z, 2018-08-16T17:26:38.102Z
#> 17 58, 17, 2016-0023, 2016-0020, 2016-07-20, 2016-06-20, 17, 17, NA, NA, 2018-08-16T17:26:44.692Z, 2018-08-16T17:26:38.316Z, 2018-08-16T17:26:44.692Z, 2018-08-16T17:26:38.316Z
#> 18 59, 18, 2016-0029, 2016-0026, 2016-07-28, 2016-07-11, 18, 18, NA, NA, 2018-08-16T17:26:44.826Z, 2018-08-16T17:26:38.641Z, 2018-08-16T17:26:44.826Z, 2018-08-16T17:26:38.641Z
#> 19 60, 19, 2016-0005, 2016-0002, 2016-07-21, 2016-06-21, 19, 19, NA, NA, 2018-08-16T17:26:44.957Z, 2018-08-16T17:26:38.873Z, 2018-08-16T17:26:44.957Z, 2018-08-16T17:26:38.873Z
#> 20                                                                                        21, 2017-0044, 2017-07-11, 20, NA, 2018-08-16T17:26:39.348Z, 2018-08-16T17:26:39.348Z
#> 21 63, 23, 2017-0023, 2017-0020, 2017-07-10, 2017-06-20, 21, 21, NA, NA, 2018-08-16T17:26:45.354Z, 2018-08-16T17:26:39.610Z, 2018-08-16T17:26:45.354Z, 2018-08-16T17:26:39.610Z
#> 22 65, 25, 2017-0005, 2017-0002, 2017-06-29, 2017-06-08, 22, 22, NA, NA, 2018-08-16T17:26:45.610Z, 2018-08-16T17:26:39.862Z, 2018-08-16T17:26:45.610Z, 2018-08-16T17:26:39.862Z
#> 23 66, 26, 2017-0011, 2017-0008, 2017-07-10, 2017-06-20, 23, 23, NA, NA, 2018-08-16T17:26:45.738Z, 2018-08-16T17:26:39.990Z, 2018-08-16T17:26:45.738Z, 2018-08-16T17:26:39.990Z
#> 24                 67, 27, 2017-0035, 2017-0032, NA, NA, 24, 24, NA, NA, 2018-08-16T17:26:45.869Z, 2018-08-16T17:26:40.118Z, 2018-08-16T17:26:45.869Z, 2018-08-16T17:26:40.118Z
#> 25 61, 20, 2017-0041, 2017-0038, 2017-08-21, 2017-07-28, 25, 25, NA, NA, 2018-08-16T17:26:45.090Z, 2018-08-16T17:26:39.194Z, 2018-08-16T17:26:45.090Z, 2018-08-16T17:26:39.194Z
#> 26 62, 22, 2017-0029, 2017-0026, 2017-07-11, 2017-06-15, 26, 26, NA, NA, 2018-08-16T17:26:45.215Z, 2018-08-16T17:26:39.483Z, 2018-08-16T17:26:45.215Z, 2018-08-16T17:26:39.483Z
#> 27 64, 24, 2017-0017, 2017-0014, 2017-07-11, 2017-06-14, 27, 27, NA, NA, 2018-08-16T17:26:45.482Z, 2018-08-16T17:26:39.739Z, 2018-08-16T17:26:45.482Z, 2018-08-16T17:26:39.739Z
#> 28 68, 28, 2017-0050, 2017-0047, 2017-09-15, 2017-09-02, 28, 28, NA, NA, 2018-08-16T17:26:45.995Z, 2018-08-16T17:26:40.244Z, 2018-08-16T17:26:45.995Z, 2018-08-16T17:26:40.244Z
#> 29 69, 29, 2016-0012, 2016-0009, 2016-07-12, 2016-06-10, 29, 29, NA, NA, 2018-08-16T17:26:46.124Z, 2018-08-16T17:26:40.376Z, 2018-08-16T17:26:46.124Z, 2018-08-16T17:26:40.376Z
#> 30 70, 30, 2016-0018, 2016-0015, 2016-07-25, 2016-06-28, 30, 30, NA, NA, 2018-08-16T17:26:46.254Z, 2018-08-16T17:26:40.511Z, 2018-08-16T17:26:46.254Z, 2018-08-16T17:26:40.511Z
#> 31 71, 31, 2016-0024, 2016-0021, 2016-07-20, 2016-06-20, 31, 31, NA, NA, 2018-08-16T17:26:46.427Z, 2018-08-16T17:26:40.636Z, 2018-08-16T17:26:46.427Z, 2018-08-16T17:26:40.636Z
#> 32 72, 32, 2016-0030, 2016-0027, 2016-07-28, 2016-07-11, 32, 32, NA, NA, 2018-08-16T17:26:46.556Z, 2018-08-16T17:26:40.761Z, 2018-08-16T17:26:46.556Z, 2018-08-16T17:26:40.761Z
#> 33 73, 33, 2016-0006, 2016-0003, 2016-07-21, 2016-06-21, 33, 33, NA, NA, 2018-08-16T17:26:46.689Z, 2018-08-16T17:26:40.930Z, 2018-08-16T17:26:46.689Z, 2018-08-16T17:26:40.930Z
#> 34                                                                                        35, 2017-0045, 2017-07-11, 34, NA, 2018-08-16T17:26:41.183Z, 2018-08-16T17:26:41.183Z
#> 35 76, 37, 2017-0024, 2017-0021, 2017-07-10, 2017-06-20, 35, 35, NA, NA, 2018-08-16T17:26:47.187Z, 2018-08-16T17:26:41.441Z, 2018-08-16T17:26:47.187Z, 2018-08-16T17:26:41.441Z
#> 36 78, 39, 2017-0006, 2017-0003, 2017-06-29, 2017-06-08, 36, 36, NA, NA, 2018-08-16T17:26:47.644Z, 2018-08-16T17:26:41.692Z, 2018-08-16T17:26:47.644Z, 2018-08-16T17:26:41.692Z
#> 37 79, 40, 2017-0012, 2017-0009, 2017-07-10, 2017-06-20, 37, 37, NA, NA, 2018-08-16T17:26:47.834Z, 2018-08-16T17:26:41.813Z, 2018-08-16T17:26:47.834Z, 2018-08-16T17:26:41.813Z
#> 38                 80, 41, 2017-0036, 2017-0033, NA, NA, 38, 38, NA, NA, 2018-08-16T17:26:47.966Z, 2018-08-16T17:26:41.937Z, 2018-08-16T17:26:47.966Z, 2018-08-16T17:26:41.937Z
#> 39 74, 34, 2017-0042, 2017-0039, 2017-08-21, 2017-07-28, 39, 39, NA, NA, 2018-08-16T17:26:46.818Z, 2018-08-16T17:26:41.063Z, 2018-08-16T17:26:46.818Z, 2018-08-16T17:26:41.063Z
#> 40 75, 36, 2017-0030, 2017-0027, 2017-07-11, 2017-06-15, 40, 40, NA, NA, 2018-08-16T17:26:46.987Z, 2018-08-16T17:26:41.310Z, 2018-08-16T17:26:46.987Z, 2018-08-16T17:26:41.310Z
#> 41 77, 38, 2017-0018, 2017-0015, 2017-07-11, 2017-06-14, 41, 41, NA, NA, 2018-08-16T17:26:47.368Z, 2018-08-16T17:26:41.570Z, 2018-08-16T17:26:47.368Z, 2018-08-16T17:26:41.570Z
#> 42 81, 42, 2017-0051, 2017-0048, 2017-09-15, 2017-09-02, 42, 42, NA, NA, 2018-08-16T17:26:48.092Z, 2018-08-16T17:26:42.064Z, 2018-08-16T17:26:48.092Z, 2018-08-16T17:26:42.064Z
#> 
#> $response
#> Response [http://localhost:3001/api/v1/traps?page=0]
#>   Date: 2018-08-16 18:40
#>   Status: 200
#>   Content-Type: application/json; charset=utf-8
#>   Size: 34.6 kB
#> 
#> 
#> attr(,"class")
#> [1] "getSuccess"
#> 
#> 
#> attr(,"class")
#> [1] "coleoGetResp"